Type Definition rp2040::adc::cs::W

type W = W<u32, CS>;

Writer for register CS

Implementations

impl W

pub fn rrobin(&mut self) -> RROBIN_W<'_>

Bits 16:20 - Round-robin sampling. 1 bit per channel. Set all bits to 0 to disable.\n Otherwise, the ADC will cycle through each enabled channel in a round-robin fashion.\n The first channel to be sampled will be the one currently indicated by AINSEL.\n AINSEL will be updated after each conversion with the newly-selected channel.

pub fn ainsel(&mut self) -> AINSEL_W<'_>

Bits 12:14 - Select analog mux input. Updated automatically in round-robin mode.

pub fn err_sticky(&mut self) -> ERR_STICKY_W<'_>

Bit 10 - Some past ADC conversion encountered an error. Write 1 to clear.

pub fn start_many(&mut self) -> START_MANY_W<'_>

Bit 3 - Continuously perform conversions whilst this bit is 1. A new conversion will start immediately after the previous finishes.

pub fn start_once(&mut self) -> START_ONCE_W<'_>

Bit 2 - Start a single conversion. Self-clearing. Ignored if start_many is asserted.

pub fn ts_en(&mut self) -> TS_EN_W<'_>

Bit 1 - Power on temperature sensor. 1 - enabled. 0 - disabled.

pub fn en(&mut self) -> EN_W<'_>

Bit 0 - Power on ADC and enable its clock.\n 1 - enabled. 0 - disabled.