Type Definition rp2040::clocks::clk_adc_ctrl::R

type R = R<u32, CLK_ADC_CTRL>;

Reader of register CLK_ADC_CTRL

Implementations

impl R

pub fn nudge(&self) -> NUDGE_R

Bit 20 - An edge on this signal shifts the phase of the output by 1 cycle of the input clock\n This can be done at any time

pub fn phase(&self) -> PHASE_R

Bits 16:17 - This delays the enable signal by up to 3 cycles of the input clock\n This must be set before the clock is enabled to have any effect

pub fn enable(&self) -> ENABLE_R

Bit 11 - Starts and stops the clock generator cleanly

pub fn kill(&self) -> KILL_R

Bit 10 - Asynchronously kills the clock generator

pub fn auxsrc(&self) -> AUXSRC_R

Bits 5:7 - Selects the auxiliary clock source, will glitch when switching