Type Definition rp2040::clocks::clk_gpout2_ctrl::W

type W = W<u32, CLK_GPOUT2_CTRL>;

Writer for register CLK_GPOUT2_CTRL

Implementations

impl W

pub fn nudge(&mut self) -> NUDGE_W<'_>

Bit 20 - An edge on this signal shifts the phase of the output by 1 cycle of the input clock\n This can be done at any time

pub fn phase(&mut self) -> PHASE_W<'_>

Bits 16:17 - This delays the enable signal by up to 3 cycles of the input clock\n This must be set before the clock is enabled to have any effect

pub fn dc50(&mut self) -> DC50_W<'_>

Bit 12 - Enables duty cycle correction for odd divisors

pub fn enable(&mut self) -> ENABLE_W<'_>

Bit 11 - Starts and stops the clock generator cleanly

pub fn kill(&mut self) -> KILL_W<'_>

Bit 10 - Asynchronously kills the clock generator

pub fn auxsrc(&mut self) -> AUXSRC_W<'_>

Bits 5:8 - Selects the auxiliary clock source, will glitch when switching