Type Definition rp2040::clocks::clk_gpout3_ctrl::AUXSRC_R

type AUXSRC_R = R<u8, AUXSRC_A>;

Reader of field AUXSRC

Implementations

impl AUXSRC_R

pub fn variant(&self) -> Variant<u8, AUXSRC_A>

Get enumerated values variant

pub fn is_clksrc_pll_sys(&self) -> bool

Checks if the value of the field is CLKSRC_PLL_SYS

pub fn is_clksrc_gpin0(&self) -> bool

Checks if the value of the field is CLKSRC_GPIN0

pub fn is_clksrc_gpin1(&self) -> bool

Checks if the value of the field is CLKSRC_GPIN1

pub fn is_clksrc_pll_usb(&self) -> bool

Checks if the value of the field is CLKSRC_PLL_USB

pub fn is_rosc_clksrc_ph(&self) -> bool

Checks if the value of the field is ROSC_CLKSRC_PH

pub fn is_xosc_clksrc(&self) -> bool

Checks if the value of the field is XOSC_CLKSRC

pub fn is_clk_sys(&self) -> bool

Checks if the value of the field is CLK_SYS

pub fn is_clk_usb(&self) -> bool

Checks if the value of the field is CLK_USB

pub fn is_clk_adc(&self) -> bool

Checks if the value of the field is CLK_ADC

pub fn is_clk_rtc(&self) -> bool

Checks if the value of the field is CLK_RTC

pub fn is_clk_ref(&self) -> bool

Checks if the value of the field is CLK_REF