Type Definition rp2040::clocks::fc0_src::FC0_SRC_R

type FC0_SRC_R = R<u8, FC0_SRC_A>;

Reader of field FC0_SRC

Implementations

impl FC0_SRC_R

pub fn variant(&self) -> Variant<u8, FC0_SRC_A>

Get enumerated values variant

pub fn is_null(&self) -> bool

Checks if the value of the field is NULL

pub fn is_pll_sys_clksrc_primary(&self) -> bool

Checks if the value of the field is PLL_SYS_CLKSRC_PRIMARY

pub fn is_pll_usb_clksrc_primary(&self) -> bool

Checks if the value of the field is PLL_USB_CLKSRC_PRIMARY

pub fn is_rosc_clksrc(&self) -> bool

Checks if the value of the field is ROSC_CLKSRC

pub fn is_rosc_clksrc_ph(&self) -> bool

Checks if the value of the field is ROSC_CLKSRC_PH

pub fn is_xosc_clksrc(&self) -> bool

Checks if the value of the field is XOSC_CLKSRC

pub fn is_clksrc_gpin0(&self) -> bool

Checks if the value of the field is CLKSRC_GPIN0

pub fn is_clksrc_gpin1(&self) -> bool

Checks if the value of the field is CLKSRC_GPIN1

pub fn is_clk_ref(&self) -> bool

Checks if the value of the field is CLK_REF

pub fn is_clk_sys(&self) -> bool

Checks if the value of the field is CLK_SYS

pub fn is_clk_peri(&self) -> bool

Checks if the value of the field is CLK_PERI

pub fn is_clk_usb(&self) -> bool

Checks if the value of the field is CLK_USB

pub fn is_clk_adc(&self) -> bool

Checks if the value of the field is CLK_ADC

pub fn is_clk_rtc(&self) -> bool

Checks if the value of the field is CLK_RTC