Type Definition rp2040::clocks::fc0_status::R

type R = R<u32, FC0_STATUS>;

Reader of register FC0_STATUS

Implementations

impl R

pub fn died(&self) -> DIED_R

Bit 28 - Test clock stopped during test

pub fn fast(&self) -> FAST_R

Bit 24 - Test clock faster than expected, only valid when status_done=1

pub fn slow(&self) -> SLOW_R

Bit 20 - Test clock slower than expected, only valid when status_done=1

pub fn fail(&self) -> FAIL_R

Bit 16 - Test failed

pub fn waiting(&self) -> WAITING_R

Bit 12 - Waiting for test clock to start

pub fn running(&self) -> RUNNING_R

Bit 8 - Test running

pub fn done(&self) -> DONE_R

Bit 4 - Test complete

pub fn pass(&self) -> PASS_R

Bit 0 - Test passed