Struct rp2040::clocks::RegisterBlock

#[repr(C)]pub struct RegisterBlock {
    pub clk_gpout0_ctrl: CLK_GPOUT0_CTRL,
    pub clk_gpout0_div: CLK_GPOUT0_DIV,
    pub clk_gpout0_selected: CLK_GPOUT0_SELECTED,
    pub clk_gpout1_ctrl: CLK_GPOUT1_CTRL,
    pub clk_gpout1_div: CLK_GPOUT1_DIV,
    pub clk_gpout1_selected: CLK_GPOUT1_SELECTED,
    pub clk_gpout2_ctrl: CLK_GPOUT2_CTRL,
    pub clk_gpout2_div: CLK_GPOUT2_DIV,
    pub clk_gpout2_selected: CLK_GPOUT2_SELECTED,
    pub clk_gpout3_ctrl: CLK_GPOUT3_CTRL,
    pub clk_gpout3_div: CLK_GPOUT3_DIV,
    pub clk_gpout3_selected: CLK_GPOUT3_SELECTED,
    pub clk_ref_ctrl: CLK_REF_CTRL,
    pub clk_ref_div: CLK_REF_DIV,
    pub clk_ref_selected: CLK_REF_SELECTED,
    pub clk_sys_ctrl: CLK_SYS_CTRL,
    pub clk_sys_div: CLK_SYS_DIV,
    pub clk_sys_selected: CLK_SYS_SELECTED,
    pub clk_peri_ctrl: CLK_PERI_CTRL,
    pub clk_peri_selected: CLK_PERI_SELECTED,
    pub clk_usb_ctrl: CLK_USB_CTRL,
    pub clk_usb_div: CLK_USB_DIV,
    pub clk_usb_selected: CLK_USB_SELECTED,
    pub clk_adc_ctrl: CLK_ADC_CTRL,
    pub clk_adc_div: CLK_ADC_DIV,
    pub clk_adc_selected: CLK_ADC_SELECTED,
    pub clk_rtc_ctrl: CLK_RTC_CTRL,
    pub clk_rtc_div: CLK_RTC_DIV,
    pub clk_rtc_selected: CLK_RTC_SELECTED,
    pub clk_sys_resus_ctrl: CLK_SYS_RESUS_CTRL,
    pub clk_sys_resus_status: CLK_SYS_RESUS_STATUS,
    pub fc0_ref_khz: FC0_REF_KHZ,
    pub fc0_min_khz: FC0_MIN_KHZ,
    pub fc0_max_khz: FC0_MAX_KHZ,
    pub fc0_delay: FC0_DELAY,
    pub fc0_interval: FC0_INTERVAL,
    pub fc0_src: FC0_SRC,
    pub fc0_status: FC0_STATUS,
    pub fc0_result: FC0_RESULT,
    pub wake_en0: WAKE_EN0,
    pub wake_en1: WAKE_EN1,
    pub sleep_en0: SLEEP_EN0,
    pub sleep_en1: SLEEP_EN1,
    pub enabled0: ENABLED0,
    pub enabled1: ENABLED1,
    pub intr: INTR,
    pub inte: INTE,
    pub intf: INTF,
    pub ints: INTS,
    // some fields omitted
}

Register block

Fields
clk_gpout0_ctrl: CLK_GPOUT0_CTRL

0x00 - Clock control, can be changed on-the-fly (except for auxsrc)

clk_gpout0_div: CLK_GPOUT0_DIV

0x04 - Clock divisor, can be changed on-the-fly

clk_gpout0_selected: CLK_GPOUT0_SELECTED

0x08 - Indicates which src is currently selected (one-hot)

clk_gpout1_ctrl: CLK_GPOUT1_CTRL

0x0c - Clock control, can be changed on-the-fly (except for auxsrc)

clk_gpout1_div: CLK_GPOUT1_DIV

0x10 - Clock divisor, can be changed on-the-fly

clk_gpout1_selected: CLK_GPOUT1_SELECTED

0x14 - Indicates which src is currently selected (one-hot)

clk_gpout2_ctrl: CLK_GPOUT2_CTRL

0x18 - Clock control, can be changed on-the-fly (except for auxsrc)

clk_gpout2_div: CLK_GPOUT2_DIV

0x1c - Clock divisor, can be changed on-the-fly

clk_gpout2_selected: CLK_GPOUT2_SELECTED

0x20 - Indicates which src is currently selected (one-hot)

clk_gpout3_ctrl: CLK_GPOUT3_CTRL

0x24 - Clock control, can be changed on-the-fly (except for auxsrc)

clk_gpout3_div: CLK_GPOUT3_DIV

0x28 - Clock divisor, can be changed on-the-fly

clk_gpout3_selected: CLK_GPOUT3_SELECTED

0x2c - Indicates which src is currently selected (one-hot)

clk_ref_ctrl: CLK_REF_CTRL

0x30 - Clock control, can be changed on-the-fly (except for auxsrc)

clk_ref_div: CLK_REF_DIV

0x34 - Clock divisor, can be changed on-the-fly

clk_ref_selected: CLK_REF_SELECTED

0x38 - Indicates which src is currently selected (one-hot)

clk_sys_ctrl: CLK_SYS_CTRL

0x3c - Clock control, can be changed on-the-fly (except for auxsrc)

clk_sys_div: CLK_SYS_DIV

0x40 - Clock divisor, can be changed on-the-fly

clk_sys_selected: CLK_SYS_SELECTED

0x44 - Indicates which src is currently selected (one-hot)

clk_peri_ctrl: CLK_PERI_CTRL

0x48 - Clock control, can be changed on-the-fly (except for auxsrc)

clk_peri_selected: CLK_PERI_SELECTED

0x50 - Indicates which src is currently selected (one-hot)

clk_usb_ctrl: CLK_USB_CTRL

0x54 - Clock control, can be changed on-the-fly (except for auxsrc)

clk_usb_div: CLK_USB_DIV

0x58 - Clock divisor, can be changed on-the-fly

clk_usb_selected: CLK_USB_SELECTED

0x5c - Indicates which src is currently selected (one-hot)

clk_adc_ctrl: CLK_ADC_CTRL

0x60 - Clock control, can be changed on-the-fly (except for auxsrc)

clk_adc_div: CLK_ADC_DIV

0x64 - Clock divisor, can be changed on-the-fly

clk_adc_selected: CLK_ADC_SELECTED

0x68 - Indicates which src is currently selected (one-hot)

clk_rtc_ctrl: CLK_RTC_CTRL

0x6c - Clock control, can be changed on-the-fly (except for auxsrc)

clk_rtc_div: CLK_RTC_DIV

0x70 - Clock divisor, can be changed on-the-fly

clk_rtc_selected: CLK_RTC_SELECTED

0x74 - Indicates which src is currently selected (one-hot)

clk_sys_resus_ctrl: CLK_SYS_RESUS_CTRL

0x78 -

clk_sys_resus_status: CLK_SYS_RESUS_STATUS

0x7c -

fc0_ref_khz: FC0_REF_KHZ

0x80 - Reference clock frequency in kHz

fc0_min_khz: FC0_MIN_KHZ

0x84 - Minimum pass frequency in kHz. This is optional. Set to 0 if you are not using the pass/fail flags

fc0_max_khz: FC0_MAX_KHZ

0x88 - Maximum pass frequency in kHz. This is optional. Set to 0x1ffffff if you are not using the pass/fail flags

fc0_delay: FC0_DELAY

0x8c - Delays the start of frequency counting to allow the mux to settle\n Delay is measured in multiples of the reference clock period

fc0_interval: FC0_INTERVAL

0x90 - The test interval is 0.98us * 2interval, but let’s call it 1us * 2interval\n The default gives a test interval of 250us

fc0_src: FC0_SRC

0x94 - Clock sent to frequency counter, set to 0 when not required\n Writing to this register initiates the frequency count

fc0_status: FC0_STATUS

0x98 - Frequency counter status

fc0_result: FC0_RESULT

0x9c - Result of frequency measurement, only valid when status_done=1

wake_en0: WAKE_EN0

0xa0 - enable clock in wake mode

wake_en1: WAKE_EN1

0xa4 - enable clock in wake mode

sleep_en0: SLEEP_EN0

0xa8 - enable clock in sleep mode

sleep_en1: SLEEP_EN1

0xac - enable clock in sleep mode

enabled0: ENABLED0

0xb0 - indicates the state of the clock enable

enabled1: ENABLED1

0xb4 - indicates the state of the clock enable

intr: INTR

0xb8 - Raw Interrupts

inte: INTE

0xbc - Interrupt Enable

intf: INTF

0xc0 - Interrupt Force

ints: INTS

0xc4 - Interrupt status after masking & forcing