Type Definition rp2040::dma::ch7_ctrl_trig::TREQ_SEL_R

type TREQ_SEL_R = R<u8, TREQ_SEL_A>;

Reader of field TREQ_SEL

Implementations

impl TREQ_SEL_R

pub fn variant(&self) -> Variant<u8, TREQ_SEL_A>

Get enumerated values variant

pub fn is_timer0(&self) -> bool

Checks if the value of the field is TIMER0

pub fn is_timer1(&self) -> bool

Checks if the value of the field is TIMER1

pub fn is_timer2(&self) -> bool

Checks if the value of the field is TIMER2

pub fn is_timer3(&self) -> bool

Checks if the value of the field is TIMER3

pub fn is_permanent(&self) -> bool

Checks if the value of the field is PERMANENT