Type Definition rp2040::dma::sniff_ctrl::W

type W = W<u32, SNIFF_CTRL>;

Writer for register SNIFF_CTRL

Implementations

impl W

pub fn out_inv(&mut self) -> OUT_INV_W<'_>

Bit 11 - If set, the result appears inverted (bitwise complement) when read. This does not affect the way the checksum is calculated; the result is transformed on-the-fly between the result register and the bus.

pub fn out_rev(&mut self) -> OUT_REV_W<'_>

Bit 10 - If set, the result appears bit-reversed when read. This does not affect the way the checksum is calculated; the result is transformed on-the-fly between the result register and the bus.

pub fn bswap(&mut self) -> BSWAP_W<'_>

Bit 9 - Locally perform a byte reverse on the sniffed data, before feeding into checksum.\n\n Note that the sniff hardware is downstream of the DMA channel byteswap performed in the read master: if channel CTRL_BSWAP and SNIFF_CTRL_BSWAP are both enabled, their effects cancel from the sniffer’s point of view.

pub fn calc(&mut self) -> CALC_W<'_>

Bits 5:8

pub fn dmach(&mut self) -> DMACH_W<'_>

Bits 1:4 - DMA channel for Sniffer to observe

pub fn en(&mut self) -> EN_W<'_>

Bit 0 - Enable sniffer