Type Definition rp2040::i2c0::ic_clr_activity::R

type R = R<u32, IC_CLR_ACTIVITY>;

Reader of register IC_CLR_ACTIVITY

Implementations

impl R

pub fn clr_activity(&self) -> CLR_ACTIVITY_R

Bit 0 - Reading this register clears the ACTIVITY interrupt if the I2C is not active anymore. If the I2C module is still active on the bus, the ACTIVITY interrupt bit continues to be set. It is automatically cleared by hardware if the module is disabled and if there is no further activity on the bus. The value read from this register to get status of the ACTIVITY interrupt (bit 8) of the IC_RAW_INTR_STAT register.\n\n Reset value: 0x0