Module rp2040::i2c0::ic_fs_spklen

I2C SS, FS or FM+ spike suppression limit\n\n This register is used to store the duration, measured in ic_clk cycles, of the longest spike that is filtered out by the spike suppression logic when the component is operating in SS, FS or FM+ modes. The relevant I2C requirement is tSP (table 4) as detailed in the I2C Bus Specification. This register must be programmed with a minimum value of 1.

Structs

IC_FS_SPKLEN_W
Write proxy for field IC_FS_SPKLEN

Type Definitions

IC_FS_SPKLEN_R
Reader of field IC_FS_SPKLEN

R
Reader of register IC_FS_SPKLEN

W
Writer for register IC_FS_SPKLEN