Module rp2040::i2c0::ic_intr_stat

I2C Interrupt Status Register\n\n Each bit in this register has a corresponding mask bit in the IC_INTR_MASK register. These bits are cleared by reading the matching interrupt clear register. The unmasked raw versions of these bits are available in the IC_RAW_INTR_STAT register.

Enums

R_ACTIVITY_A
See IC_RAW_INTR_STAT for a detailed description of R_ACTIVITY bit.\n\n Reset value: 0x0

R_GEN_CALL_A
See IC_RAW_INTR_STAT for a detailed description of R_GEN_CALL bit.\n\n Reset value: 0x0

R_MASTER_ON_HOLD_A
See IC_RAW_INTR_STAT for a detailed description of R_MASTER_ON_HOLD bit.\n\n Reset value: 0x0

R_RD_REQ_A
See IC_RAW_INTR_STAT for a detailed description of R_RD_REQ bit.\n\n Reset value: 0x0

R_RESTART_DET_A
See IC_RAW_INTR_STAT for a detailed description of R_RESTART_DET bit.\n\n Reset value: 0x0

R_RX_DONE_A
See IC_RAW_INTR_STAT for a detailed description of R_RX_DONE bit.\n\n Reset value: 0x0

R_RX_FULL_A
See IC_RAW_INTR_STAT for a detailed description of R_RX_FULL bit.\n\n Reset value: 0x0

R_RX_OVER_A
See IC_RAW_INTR_STAT for a detailed description of R_RX_OVER bit.\n\n Reset value: 0x0

R_RX_UNDER_A
See IC_RAW_INTR_STAT for a detailed description of R_RX_UNDER bit.\n\n Reset value: 0x0

R_START_DET_A
See IC_RAW_INTR_STAT for a detailed description of R_START_DET bit.\n\n Reset value: 0x0

R_STOP_DET_A
See IC_RAW_INTR_STAT for a detailed description of R_STOP_DET bit.\n\n Reset value: 0x0

R_TX_ABRT_A
See IC_RAW_INTR_STAT for a detailed description of R_TX_ABRT bit.\n\n Reset value: 0x0

R_TX_EMPTY_A
See IC_RAW_INTR_STAT for a detailed description of R_TX_EMPTY bit.\n\n Reset value: 0x0

R_TX_OVER_A
See IC_RAW_INTR_STAT for a detailed description of R_TX_OVER bit.\n\n Reset value: 0x0

Type Definitions

R
Reader of register IC_INTR_STAT

R_ACTIVITY_R
Reader of field R_ACTIVITY

R_GEN_CALL_R
Reader of field R_GEN_CALL

R_MASTER_ON_HOLD_R
Reader of field R_MASTER_ON_HOLD

R_RD_REQ_R
Reader of field R_RD_REQ

R_RESTART_DET_R
Reader of field R_RESTART_DET

R_RX_DONE_R
Reader of field R_RX_DONE

R_RX_FULL_R
Reader of field R_RX_FULL

R_RX_OVER_R
Reader of field R_RX_OVER

R_RX_UNDER_R
Reader of field R_RX_UNDER

R_START_DET_R
Reader of field R_START_DET

R_STOP_DET_R
Reader of field R_STOP_DET

R_TX_ABRT_R
Reader of field R_TX_ABRT

R_TX_EMPTY_R
Reader of field R_TX_EMPTY

R_TX_OVER_R
Reader of field R_TX_OVER