Type Definition rp2040::i2c0::ic_rx_tl::R

type R = R<u32, IC_RX_TL>;

Reader of register IC_RX_TL

Implementations

impl R

pub fn rx_tl(&self) -> RX_TL_R

Bits 0:7 - Receive FIFO Threshold Level.\n\n Controls the level of entries (or above) that triggers the RX_FULL interrupt (bit 2 in IC_RAW_INTR_STAT register). The valid range is 0-255, with the additional restriction that hardware does not allow this value to be set to a value larger than the depth of the buffer. If an attempt is made to do that, the actual value set will be the maximum depth of the buffer. A value of 0 sets the threshold for 1 entry, and a value of 255 sets the threshold for 256 entries.