Module rp2040::i2c0::ic_tar

I2C Target Address Register\n\n This register is 12 bits wide, and bits 31:12 are reserved. This register can be written to only when IC_ENABLE[0] is set to 0.\n\n Note: If the software or application is aware that the DW_apb_i2c is not using the TAR address for the pending commands in the Tx FIFO, then it is possible to update the TAR address even while the Tx FIFO has entries (IC_STATUS[2]= 0). - It is not necessary to perform any write to this register if DW_apb_i2c is enabled as an I2C slave only.

Structs

GC_OR_START_W
Write proxy for field GC_OR_START

IC_TAR_W
Write proxy for field IC_TAR

SPECIAL_W
Write proxy for field SPECIAL

Enums

GC_OR_START_A
If bit 11 (SPECIAL) is set to 1 and bit 13(Device-ID) is set to 0, then this bit indicates whether a General Call or START byte command is to be performed by the DW_apb_i2c. - 0: General Call Address - after issuing a General Call, only writes may be performed. Attempting to issue a read command results in setting bit 6 (TX_ABRT) of the IC_RAW_INTR_STAT register. The DW_apb_i2c remains in General Call mode until the SPECIAL bit value (bit 11) is cleared. - 1: START BYTE Reset value: 0x0

SPECIAL_A
This bit indicates whether software performs a Device-ID or General Call or START BYTE command. - 0: ignore bit 10 GC_OR_START and use IC_TAR normally - 1: perform special I2C command as specified in Device_ID or GC_OR_START bit Reset value: 0x0

Type Definitions

GC_OR_START_R
Reader of field GC_OR_START

IC_TAR_R
Reader of field IC_TAR

R
Reader of register IC_TAR

SPECIAL_R
Reader of field SPECIAL

W
Writer for register IC_TAR