Module rp2040::i2c0::ic_tx_abrt_source

I2C Transmit Abort Source Register\n\n This register has 32 bits that indicate the source of the TX_ABRT bit. Except for Bit 9, this register is cleared whenever the IC_CLR_TX_ABRT register or the IC_CLR_INTR register is read. To clear Bit 9, the source of the ABRT_SBYTE_NORSTRT must be fixed first; RESTART must be enabled (IC_CON[5]=1), the SPECIAL bit must be cleared (IC_TAR[11]), or the GC_OR_START bit must be cleared (IC_TAR[10]).\n\n Once the source of the ABRT_SBYTE_NORSTRT is fixed, then this bit can be cleared in the same manner as other bits in this register. If the source of the ABRT_SBYTE_NORSTRT is not fixed before attempting to clear this bit, Bit 9 clears for one cycle and is then re-asserted.

Enums

ABRT_7B_ADDR_NOACK_A
This field indicates that the Master is in 7-bit addressing mode and the address sent was not acknowledged by any slave.\n\n Reset value: 0x0\n\n Role of DW_apb_i2c: Master-Transmitter or Master-Receiver

ABRT_10ADDR1_NOACK_A
This field indicates that the Master is in 10-bit address mode and the first 10-bit address byte was not acknowledged by any slave.\n\n Reset value: 0x0\n\n Role of DW_apb_i2c: Master-Transmitter or Master-Receiver

ABRT_10ADDR2_NOACK_A
This field indicates that the Master is in 10-bit address mode and that the second address byte of the 10-bit address was not acknowledged by any slave.\n\n Reset value: 0x0\n\n Role of DW_apb_i2c: Master-Transmitter or Master-Receiver

ABRT_10B_RD_NORSTRT_A
This field indicates that the restart is disabled (IC_RESTART_EN bit (IC_CON[5]) =0) and the master sends a read command in 10-bit addressing mode.\n\n Reset value: 0x0\n\n Role of DW_apb_i2c: Master-Receiver

ABRT_GCALL_NOACK_A
This field indicates that DW_apb_i2c in master mode has sent a General Call and no slave on the bus acknowledged the General Call.\n\n Reset value: 0x0\n\n Role of DW_apb_i2c: Master-Transmitter

ABRT_GCALL_READ_A
This field indicates that DW_apb_i2c in the master mode has sent a General Call but the user programmed the byte following the General Call to be a read from the bus (IC_DATA_CMD[9] is set to 1).\n\n Reset value: 0x0\n\n Role of DW_apb_i2c: Master-Transmitter

ABRT_HS_ACKDET_A
This field indicates that the Master is in High Speed mode and the High Speed Master code was acknowledged (wrong behavior).\n\n Reset value: 0x0\n\n Role of DW_apb_i2c: Master

ABRT_HS_NORSTRT_A
This field indicates that the restart is disabled (IC_RESTART_EN bit (IC_CON[5]) =0) and the user is trying to use the master to transfer data in High Speed mode.\n\n Reset value: 0x0\n\n Role of DW_apb_i2c: Master-Transmitter or Master-Receiver

ABRT_MASTER_DIS_A
This field indicates that the User tries to initiate a Master operation with the Master mode disabled.\n\n Reset value: 0x0\n\n Role of DW_apb_i2c: Master-Transmitter or Master-Receiver

ABRT_SBYTE_ACKDET_A
This field indicates that the Master has sent a START Byte and the START Byte was acknowledged (wrong behavior).\n\n Reset value: 0x0\n\n Role of DW_apb_i2c: Master

ABRT_SBYTE_NORSTRT_A
To clear Bit 9, the source of the ABRT_SBYTE_NORSTRT must be fixed first; restart must be enabled (IC_CON[5]=1), the SPECIAL bit must be cleared (IC_TAR[11]), or the GC_OR_START bit must be cleared (IC_TAR[10]). Once the source of the ABRT_SBYTE_NORSTRT is fixed, then this bit can be cleared in the same manner as other bits in this register. If the source of the ABRT_SBYTE_NORSTRT is not fixed before attempting to clear this bit, bit 9 clears for one cycle and then gets reasserted. When this field is set to 1, the restart is disabled (IC_RESTART_EN bit (IC_CON[5]) =0) and the user is trying to send a START Byte.\n\n Reset value: 0x0\n\n Role of DW_apb_i2c: Master

ABRT_SLVFLUSH_TXFIFO_A
This field specifies that the Slave has received a read command and some data exists in the TX FIFO, so the slave issues a TX_ABRT interrupt to flush old data in TX FIFO.\n\n Reset value: 0x0\n\n Role of DW_apb_i2c: Slave-Transmitter

ABRT_SLVRD_INTX_A
1: When the processor side responds to a slave mode request for data to be transmitted to a remote master and user writes a 1 in CMD (bit 8) of IC_DATA_CMD register.\n\n Reset value: 0x0\n\n Role of DW_apb_i2c: Slave-Transmitter

ABRT_SLV_ARBLOST_A
This field indicates that a Slave has lost the bus while transmitting data to a remote master. IC_TX_ABRT_SOURCE[12] is set at the same time. Note: Even though the slave never ‘owns’ the bus, something could go wrong on the bus. This is a fail safe check. For instance, during a data transmission at the low-to-high transition of SCL, if what is on the data bus is not what is supposed to be transmitted, then DW_apb_i2c no longer own the bus.\n\n Reset value: 0x0\n\n Role of DW_apb_i2c: Slave-Transmitter

ABRT_TXDATA_NOACK_A
This field indicates the master-mode only bit. When the master receives an acknowledgement for the address, but when it sends data byte(s) following the address, it did not receive an acknowledge from the remote slave(s).\n\n Reset value: 0x0\n\n Role of DW_apb_i2c: Master-Transmitter

ABRT_USER_ABRT_A
This is a master-mode-only bit. Master has detected the transfer abort (IC_ENABLE[1])\n\n Reset value: 0x0\n\n Role of DW_apb_i2c: Master-Transmitter

ARB_LOST_A
This field specifies that the Master has lost arbitration, or if IC_TX_ABRT_SOURCE[14] is also set, then the slave transmitter has lost arbitration.\n\n Reset value: 0x0\n\n Role of DW_apb_i2c: Master-Transmitter or Slave-Transmitter

Type Definitions

ABRT_7B_ADDR_NOACK_R
Reader of field ABRT_7B_ADDR_NOACK

ABRT_10ADDR1_NOACK_R
Reader of field ABRT_10ADDR1_NOACK

ABRT_10ADDR2_NOACK_R
Reader of field ABRT_10ADDR2_NOACK

ABRT_10B_RD_NORSTRT_R
Reader of field ABRT_10B_RD_NORSTRT

ABRT_GCALL_NOACK_R
Reader of field ABRT_GCALL_NOACK

ABRT_GCALL_READ_R
Reader of field ABRT_GCALL_READ

ABRT_HS_ACKDET_R
Reader of field ABRT_HS_ACKDET

ABRT_HS_NORSTRT_R
Reader of field ABRT_HS_NORSTRT

ABRT_MASTER_DIS_R
Reader of field ABRT_MASTER_DIS

ABRT_SBYTE_ACKDET_R
Reader of field ABRT_SBYTE_ACKDET

ABRT_SBYTE_NORSTRT_R
Reader of field ABRT_SBYTE_NORSTRT

ABRT_SLVFLUSH_TXFIFO_R
Reader of field ABRT_SLVFLUSH_TXFIFO

ABRT_SLVRD_INTX_R
Reader of field ABRT_SLVRD_INTX

ABRT_SLV_ARBLOST_R
Reader of field ABRT_SLV_ARBLOST

ABRT_TXDATA_NOACK_R
Reader of field ABRT_TXDATA_NOACK

ABRT_USER_ABRT_R
Reader of field ABRT_USER_ABRT

ARB_LOST_R
Reader of field ARB_LOST

R
Reader of register IC_TX_ABRT_SOURCE

TX_FLUSH_CNT_R
Reader of field TX_FLUSH_CNT