Type Definition rp2040::io_bank0::gpio14_status::R

type R = R<u32, GPIO14_STATUS>;

Reader of register GPIO14_STATUS

Implementations

impl R

pub fn irqtoproc(&self) -> IRQTOPROC_R

Bit 26 - interrupt to processors, after override is applied

pub fn irqfrompad(&self) -> IRQFROMPAD_R

Bit 24 - interrupt from pad before override is applied

pub fn intoperi(&self) -> INTOPERI_R

Bit 19 - input signal to peripheral, after override is applied

pub fn infrompad(&self) -> INFROMPAD_R

Bit 17 - input signal from pad, before override is applied

pub fn oetopad(&self) -> OETOPAD_R

Bit 13 - output enable to pad after register override is applied

pub fn oefromperi(&self) -> OEFROMPERI_R

Bit 12 - output enable from selected peripheral, before register override is applied

pub fn outtopad(&self) -> OUTTOPAD_R

Bit 9 - output signal to pad after register override is applied

pub fn outfromperi(&self) -> OUTFROMPERI_R

Bit 8 - output signal from selected peripheral, before register override is applied