Type Definition rp2040::io_bank0::gpio1_ctrl::FUNCSEL_R

type FUNCSEL_R = R<u8, FUNCSEL_A>;

Reader of field FUNCSEL

Implementations

impl FUNCSEL_R

pub fn variant(&self) -> Variant<u8, FUNCSEL_A>

Get enumerated values variant

pub fn is_jtag_tms(&self) -> bool

Checks if the value of the field is JTAG_TMS

pub fn is_spi0_ss_n(&self) -> bool

Checks if the value of the field is SPI0_SS_N

pub fn is_uart0_rx(&self) -> bool

Checks if the value of the field is UART0_RX

pub fn is_i2c0_scl(&self) -> bool

Checks if the value of the field is I2C0_SCL

pub fn is_pwm_b_0(&self) -> bool

Checks if the value of the field is PWM_B_0

pub fn is_sio_1(&self) -> bool

Checks if the value of the field is SIO_1

pub fn is_pio0_1(&self) -> bool

Checks if the value of the field is PIO0_1

pub fn is_pio1_1(&self) -> bool

Checks if the value of the field is PIO1_1

pub fn is_usb_muxing_vbus_detect(&self) -> bool

Checks if the value of the field is USB_MUXING_VBUS_DETECT

pub fn is_null(&self) -> bool

Checks if the value of the field is NULL