Type Definition rp2040::pio0::sm0_shiftctrl::R

type R = R<u32, SM0_SHIFTCTRL>;

Reader of register SM0_SHIFTCTRL

Implementations

impl R

pub fn fjoin_rx(&self) -> FJOIN_RX_R

Bit 31 - When 1, RX FIFO steals the TX FIFO’s storage, and becomes twice as deep.\n TX FIFO is disabled as a result (always reads as both full and empty).\n FIFOs are flushed when this bit is changed.

pub fn fjoin_tx(&self) -> FJOIN_TX_R

Bit 30 - When 1, TX FIFO steals the RX FIFO’s storage, and becomes twice as deep.\n RX FIFO is disabled as a result (always reads as both full and empty).\n FIFOs are flushed when this bit is changed.

pub fn pull_thresh(&self) -> PULL_THRESH_R

Bits 25:29 - Number of bits shifted out of TXSR before autopull or conditional pull.\n Write 0 for value of 32.

pub fn push_thresh(&self) -> PUSH_THRESH_R

Bits 20:24 - Number of bits shifted into RXSR before autopush or conditional push.\n Write 0 for value of 32.

pub fn out_shiftdir(&self) -> OUT_SHIFTDIR_R

Bit 19 - 1 = shift out of output shift register to right. 0 = to left.

pub fn in_shiftdir(&self) -> IN_SHIFTDIR_R

Bit 18 - 1 = shift input shift register to right (data enters from left). 0 = to left.

pub fn autopull(&self) -> AUTOPULL_R

Bit 17 - Pull automatically when the output shift register is emptied

pub fn autopush(&self) -> AUTOPUSH_R

Bit 16 - Push automatically when the input shift register is filled