Type Definition rp2040::pio0::sm1_execctrl::R

type R = R<u32, SM1_EXECCTRL>;

Reader of register SM1_EXECCTRL

Implementations

impl R

pub fn exec_stalled(&self) -> EXEC_STALLED_R

Bit 31 - An instruction written to SMx_INSTR is stalled, and latched by the\n state machine. Will clear once the instruction completes.

pub fn side_en(&self) -> SIDE_EN_R

Bit 30 - If 1, the delay MSB is used as side-set enable, rather than a\n side-set data bit. This allows instructions to perform side-set optionally,\n rather than on every instruction.

pub fn side_pindir(&self) -> SIDE_PINDIR_R

Bit 29 - Side-set data is asserted to pin OEs instead of pin values

pub fn jmp_pin(&self) -> JMP_PIN_R

Bits 24:28 - The GPIO number to use as condition for JMP PIN. Unaffected by input mapping.

pub fn out_en_sel(&self) -> OUT_EN_SEL_R

Bits 19:23 - Which data bit to use for inline OUT enable

pub fn inline_out_en(&self) -> INLINE_OUT_EN_R

Bit 18 - If 1, use a bit of OUT data as an auxiliary write enable\n When used in conjunction with OUT_STICKY, writes with an enable of 0 will\n deassert the latest pin write. This can create useful masking/override behaviour\n due to the priority ordering of state machine pin writes (SM0 < SM1 < …)

pub fn out_sticky(&self) -> OUT_STICKY_R

Bit 17 - Continuously assert the most recent OUT/SET to the pins

pub fn wrap_top(&self) -> WRAP_TOP_R

Bits 12:16 - After reaching this address, execution is wrapped to wrap_bottom.\n If the instruction is a jump, and the jump condition is true, the jump takes priority.

pub fn wrap_bottom(&self) -> WRAP_BOTTOM_R

Bits 7:11 - After reaching wrap_top, execution is wrapped to this address.

pub fn status_sel(&self) -> STATUS_SEL_R

Bit 4 - Comparison used for the MOV x, STATUS instruction.

pub fn status_n(&self) -> STATUS_N_R

Bits 0:3 - Comparison level for the MOV x, STATUS instruction