Type Definition rp2040::pio0::sm2_shiftctrl::W

type W = W<u32, SM2_SHIFTCTRL>;

Writer for register SM2_SHIFTCTRL

Implementations

impl W

pub fn fjoin_rx(&mut self) -> FJOIN_RX_W<'_>

Bit 31 - When 1, RX FIFO steals the TX FIFO’s storage, and becomes twice as deep.\n TX FIFO is disabled as a result (always reads as both full and empty).\n FIFOs are flushed when this bit is changed.

pub fn fjoin_tx(&mut self) -> FJOIN_TX_W<'_>

Bit 30 - When 1, TX FIFO steals the RX FIFO’s storage, and becomes twice as deep.\n RX FIFO is disabled as a result (always reads as both full and empty).\n FIFOs are flushed when this bit is changed.

pub fn pull_thresh(&mut self) -> PULL_THRESH_W<'_>

Bits 25:29 - Number of bits shifted out of TXSR before autopull or conditional pull.\n Write 0 for value of 32.

pub fn push_thresh(&mut self) -> PUSH_THRESH_W<'_>

Bits 20:24 - Number of bits shifted into RXSR before autopush or conditional push.\n Write 0 for value of 32.

pub fn out_shiftdir(&mut self) -> OUT_SHIFTDIR_W<'_>

Bit 19 - 1 = shift out of output shift register to right. 0 = to left.

pub fn in_shiftdir(&mut self) -> IN_SHIFTDIR_W<'_>

Bit 18 - 1 = shift input shift register to right (data enters from left). 0 = to left.

pub fn autopull(&mut self) -> AUTOPULL_W<'_>

Bit 17 - Pull automatically when the output shift register is emptied

pub fn autopush(&mut self) -> AUTOPUSH_W<'_>

Bit 16 - Push automatically when the input shift register is filled