Type Definition rp2040::pio0::sm3_execctrl::STATUS_SEL_R

type STATUS_SEL_R = R<bool, STATUS_SEL_A>;

Reader of field STATUS_SEL

Implementations

impl STATUS_SEL_R

pub fn variant(&self) -> STATUS_SEL_A

Get enumerated values variant

pub fn is_txlevel(&self) -> bool

Checks if the value of the field is TXLEVEL

pub fn is_rxlevel(&self) -> bool

Checks if the value of the field is RXLEVEL