Type Definition rp2040::ppb::cpuid::R

type R = R<u32, CPUID>;

Reader of register CPUID

Implementations

impl R

pub fn implementer(&self) -> IMPLEMENTER_R

Bits 24:31 - Implementor code: 0x41 = ARM

pub fn variant(&self) -> VARIANT_R

Bits 20:23 - Major revision number n in the rnpm revision status:\n 0x0 = Revision 0.

pub fn architecture(&self) -> ARCHITECTURE_R

Bits 16:19 - Constant that defines the architecture of the processor:\n 0xC = ARMv6-M architecture.

pub fn partno(&self) -> PARTNO_R

Bits 4:15 - Number of processor within family: 0xC60 = Cortex-M0+

pub fn revision(&self) -> REVISION_R

Bits 0:3 - Minor revision number m in the rnpm revision status:\n 0x1 = Patch 1.