Module rp2040::ppb::icsr

Use the Interrupt Control State Register to set a pending Non-Maskable Interrupt (NMI), set or clear a pending PendSV, set or clear a pending SysTick, check for pending exceptions, check the vector number of the highest priority pended exception, check the vector number of the active exception.

Structs

NMIPENDSET_W
Write proxy for field NMIPENDSET

PENDSTCLR_W
Write proxy for field PENDSTCLR

PENDSTSET_W
Write proxy for field PENDSTSET

PENDSVCLR_W
Write proxy for field PENDSVCLR

PENDSVSET_W
Write proxy for field PENDSVSET

Type Definitions

ISRPENDING_R
Reader of field ISRPENDING

ISRPREEMPT_R
Reader of field ISRPREEMPT

NMIPENDSET_R
Reader of field NMIPENDSET

PENDSTCLR_R
Reader of field PENDSTCLR

PENDSTSET_R
Reader of field PENDSTSET

PENDSVCLR_R
Reader of field PENDSVCLR

PENDSVSET_R
Reader of field PENDSVSET

R
Reader of register ICSR

VECTACTIVE_R
Reader of field VECTACTIVE

VECTPENDING_R
Reader of field VECTPENDING

W
Writer for register ICSR