Type Definition rp2040::ppb::icsr::R

type R = R<u32, ICSR>;

Reader of register ICSR

Implementations

impl R

pub fn nmipendset(&self) -> NMIPENDSET_R

Bit 31 - Setting this bit will activate an NMI. Since NMI is the highest priority exception, it will activate as soon as it is registered.\n NMI set-pending bit.\n Write:\n 0 = No effect.\n 1 = Changes NMI exception state to pending.\n Read:\n 0 = NMI exception is not pending.\n 1 = NMI exception is pending.\n Because NMI is the highest-priority exception, normally the processor enters the NMI\n exception handler as soon as it detects a write of 1 to this bit. Entering the handler then clears\n this bit to 0. This means a read of this bit by the NMI exception handler returns 1 only if the\n NMI signal is reasserted while the processor is executing that handler.

pub fn pendsvset(&self) -> PENDSVSET_R

Bit 28 - PendSV set-pending bit.\n Write:\n 0 = No effect.\n 1 = Changes PendSV exception state to pending.\n Read:\n 0 = PendSV exception is not pending.\n 1 = PendSV exception is pending.\n Writing 1 to this bit is the only way to set the PendSV exception state to pending.

pub fn pendsvclr(&self) -> PENDSVCLR_R

Bit 27 - PendSV clear-pending bit.\n Write:\n 0 = No effect.\n 1 = Removes the pending state from the PendSV exception.

pub fn pendstset(&self) -> PENDSTSET_R

Bit 26 - SysTick exception set-pending bit.\n Write:\n 0 = No effect.\n 1 = Changes SysTick exception state to pending.\n Read:\n 0 = SysTick exception is not pending.\n 1 = SysTick exception is pending.

pub fn pendstclr(&self) -> PENDSTCLR_R

Bit 25 - SysTick exception clear-pending bit.\n Write:\n 0 = No effect.\n 1 = Removes the pending state from the SysTick exception.\n This bit is WO. On a register read its value is Unknown.

pub fn isrpreempt(&self) -> ISRPREEMPT_R

Bit 23 - The system can only access this bit when the core is halted. It indicates that a pending interrupt is to be taken in the next running cycle. If C_MASKINTS is clear in the Debug Halting Control and Status Register, the interrupt is serviced.

pub fn isrpending(&self) -> ISRPENDING_R

Bit 22 - External interrupt pending flag

pub fn vectpending(&self) -> VECTPENDING_R

Bits 12:20 - Indicates the exception number for the highest priority pending exception: 0 = no pending exceptions. Non zero = The pending state includes the effect of memory-mapped enable and mask registers. It does not include the PRIMASK special-purpose register qualifier.

pub fn vectactive(&self) -> VECTACTIVE_R

Bits 0:8 - Active exception number field. Reset clears the VECTACTIVE field.