Module rp2040::ppb::mpu_ctrl

Use the MPU Control Register to enable and disable the MPU, and to control whether the default memory map is enabled as a background region for privileged accesses, and whether the MPU is enabled for HardFaults and NMIs.

Structs

ENABLE_W
Write proxy for field ENABLE

HFNMIENA_W
Write proxy for field HFNMIENA

PRIVDEFENA_W
Write proxy for field PRIVDEFENA

Type Definitions

ENABLE_R
Reader of field ENABLE

HFNMIENA_R
Reader of field HFNMIENA

PRIVDEFENA_R
Reader of field PRIVDEFENA

R
Reader of register MPU_CTRL

W
Writer for register MPU_CTRL