Type Definition rp2040::ppb::mpu_rbar::R

type R = R<u32, MPU_RBAR>;

Reader of register MPU_RBAR

Implementations

impl R

pub fn addr(&self) -> ADDR_R

Bits 8:31 - Base address of the region.

pub fn valid(&self) -> VALID_R

Bit 4 - On writes, indicates whether the write must update the base address of the region identified by the REGION field, updating the MPU_RNR to indicate this new region.\n Write:\n 0 = MPU_RNR not changed, and the processor:\n Updates the base address for the region specified in the MPU_RNR.\n Ignores the value of the REGION field.\n 1 = The processor:\n Updates the value of the MPU_RNR to the value of the REGION field.\n Updates the base address for the region specified in the REGION field.\n Always reads as zero.

pub fn region(&self) -> REGION_R

Bits 0:3 - On writes, specifies the number of the region whose base address to update provided VALID is set written as 1. On reads, returns bits [3:0] of MPU_RNR.