Module rp2040::ppb::scr

System Control Register. Use the System Control Register for power-management functions: signal to the system when the processor can enter a low power state, control how the processor enters and exits low power states.

Structs

SEVONPEND_W
Write proxy for field SEVONPEND

SLEEPDEEP_W
Write proxy for field SLEEPDEEP

SLEEPONEXIT_W
Write proxy for field SLEEPONEXIT

Type Definitions

R
Reader of register SCR

SEVONPEND_R
Reader of field SEVONPEND

SLEEPDEEP_R
Reader of field SLEEPDEEP

SLEEPONEXIT_R
Reader of field SLEEPONEXIT

W
Writer for register SCR