Module rp2040::ppb::syst_rvr

Use the SysTick Reload Value Register to specify the start value to load into the current value register when the counter reaches 0. It can be any value between 0 and 0x00FFFFFF. A start value of 0 is possible, but has no effect because the SysTick interrupt and COUNTFLAG are activated when counting from 1 to 0. The reset value of this register is UNKNOWN.\n To generate a multi-shot timer with a period of N processor clock cycles, use a RELOAD value of N-1. For example, if the SysTick interrupt is required every 100 clock pulses, set RELOAD to 99.

Structs

RELOAD_W
Write proxy for field RELOAD

Type Definitions

R
Reader of register SYST_RVR

RELOAD_R
Reader of field RELOAD

W
Writer for register SYST_RVR