Module rp2040::psm

PSM

Modules

done
Indicates the peripheral’s registers are ready to access.

frce_off
Force into reset (i.e. power it off)

frce_on
Force block out of reset (i.e. power it on)

wdsel
Set to 1 if this peripheral should be reset when the watchdog fires.

Structs

RegisterBlock
Register block

Type Definitions

DONE
Indicates the peripheral’s registers are ready to access.

FRCE_OFF
Force into reset (i.e. power it off)

FRCE_ON
Force block out of reset (i.e. power it on)

WDSEL
Set to 1 if this peripheral should be reset when the watchdog fires.