Type Definition rp2040::sio::interp0_ctrl_lane0::W

type W = W<u32, INTERP0_CTRL_LANE0>;

Writer for register INTERP0_CTRL_LANE0

Implementations

impl W

pub fn blend(&mut self) -> BLEND_W<'_>

Bit 21 - Only present on INTERP0 on each core. If BLEND mode is enabled:\n - LANE1 result is a linear interpolation between BASE0 and BASE1, controlled\n by the 8 LSBs of lane 1 shift and mask value (a fractional number between\n 0 and 255/256ths)\n - LANE0 result does not have BASE0 added (yields only the 8 LSBs of lane 1 shift+mask value)\n - FULL result does not have lane 1 shift+mask value added (BASE2 + lane 0 shift+mask)\n LANE1 SIGNED flag controls whether the interpolation is signed or unsigned.

pub fn force_msb(&mut self) -> FORCE_MSB_W<'_>

Bits 19:20 - ORed into bits 29:28 of the lane result presented to the processor on the bus.\n No effect on the internal 32-bit datapath. Handy for using a lane to generate sequence\n of pointers into flash or SRAM.

pub fn add_raw(&mut self) -> ADD_RAW_W<'_>

Bit 18 - If 1, mask + shift is bypassed for LANE0 result. This does not affect FULL result.

pub fn cross_result(&mut self) -> CROSS_RESULT_W<'_>

Bit 17 - If 1, feed the opposite lane’s result into this lane’s accumulator on POP.

pub fn cross_input(&mut self) -> CROSS_INPUT_W<'_>

Bit 16 - If 1, feed the opposite lane’s accumulator into this lane’s shift + mask hardware.\n Takes effect even if ADD_RAW is set (the CROSS_INPUT mux is before the shift+mask bypass)

pub fn signed(&mut self) -> SIGNED_W<'_>

Bit 15 - If SIGNED is set, the shifted and masked accumulator value is sign-extended to 32 bits\n before adding to BASE0, and LANE0 PEEK/POP appear extended to 32 bits when read by processor.

pub fn mask_msb(&mut self) -> MASK_MSB_W<'_>

Bits 10:14 - The most-significant bit allowed to pass by the mask (inclusive)\n Setting MSB < LSB may cause chip to turn inside-out

pub fn mask_lsb(&mut self) -> MASK_LSB_W<'_>

Bits 5:9 - The least-significant bit allowed to pass by the mask (inclusive)

pub fn shift(&mut self) -> SHIFT_W<'_>

Bits 0:4 - Logical right-shift applied to accumulator before masking