Type Definition rp2040::sio::interp1_ctrl_lane1::R

type R = R<u32, INTERP1_CTRL_LANE1>;

Reader of register INTERP1_CTRL_LANE1

Implementations

impl R

pub fn force_msb(&self) -> FORCE_MSB_R

Bits 19:20 - ORed into bits 29:28 of the lane result presented to the processor on the bus.\n No effect on the internal 32-bit datapath. Handy for using a lane to generate sequence\n of pointers into flash or SRAM.

pub fn add_raw(&self) -> ADD_RAW_R

Bit 18 - If 1, mask + shift is bypassed for LANE1 result. This does not affect FULL result.

pub fn cross_result(&self) -> CROSS_RESULT_R

Bit 17 - If 1, feed the opposite lane’s result into this lane’s accumulator on POP.

pub fn cross_input(&self) -> CROSS_INPUT_R

Bit 16 - If 1, feed the opposite lane’s accumulator into this lane’s shift + mask hardware.\n Takes effect even if ADD_RAW is set (the CROSS_INPUT mux is before the shift+mask bypass)

pub fn signed(&self) -> SIGNED_R

Bit 15 - If SIGNED is set, the shifted and masked accumulator value is sign-extended to 32 bits\n before adding to BASE1, and LANE1 PEEK/POP appear extended to 32 bits when read by processor.

pub fn mask_msb(&self) -> MASK_MSB_R

Bits 10:14 - The most-significant bit allowed to pass by the mask (inclusive)\n Setting MSB < LSB may cause chip to turn inside-out

pub fn mask_lsb(&self) -> MASK_LSB_R

Bits 5:9 - The least-significant bit allowed to pass by the mask (inclusive)

pub fn shift(&self) -> SHIFT_R

Bits 0:4 - Logical right-shift applied to accumulator before masking