Struct rp2040::sio::RegisterBlock

#[repr(C)]pub struct RegisterBlock {
    pub cpuid: CPUID,
    pub gpio_in: GPIO_IN,
    pub gpio_hi_in: GPIO_HI_IN,
    pub gpio_out: GPIO_OUT,
    pub gpio_out_set: GPIO_OUT_SET,
    pub gpio_out_clr: GPIO_OUT_CLR,
    pub gpio_out_xor: GPIO_OUT_XOR,
    pub gpio_oe: GPIO_OE,
    pub gpio_oe_set: GPIO_OE_SET,
    pub gpio_oe_clr: GPIO_OE_CLR,
    pub gpio_oe_xor: GPIO_OE_XOR,
    pub gpio_hi_out: GPIO_HI_OUT,
    pub gpio_hi_out_set: GPIO_HI_OUT_SET,
    pub gpio_hi_out_clr: GPIO_HI_OUT_CLR,
    pub gpio_hi_out_xor: GPIO_HI_OUT_XOR,
    pub gpio_hi_oe: GPIO_HI_OE,
    pub gpio_hi_oe_set: GPIO_HI_OE_SET,
    pub gpio_hi_oe_clr: GPIO_HI_OE_CLR,
    pub gpio_hi_oe_xor: GPIO_HI_OE_XOR,
    pub fifo_st: FIFO_ST,
    pub fifo_wr: FIFO_WR,
    pub fifo_rd: FIFO_RD,
    pub spinlock_st: SPINLOCK_ST,
    pub div_udividend: DIV_UDIVIDEND,
    pub div_udivisor: DIV_UDIVISOR,
    pub div_sdividend: DIV_SDIVIDEND,
    pub div_sdivisor: DIV_SDIVISOR,
    pub div_quotient: DIV_QUOTIENT,
    pub div_remainder: DIV_REMAINDER,
    pub div_csr: DIV_CSR,
    pub interp0_accum0: INTERP0_ACCUM0,
    pub interp0_accum1: INTERP0_ACCUM1,
    pub interp0_base0: INTERP0_BASE0,
    pub interp0_base1: INTERP0_BASE1,
    pub interp0_base2: INTERP0_BASE2,
    pub interp0_pop_lane0: INTERP0_POP_LANE0,
    pub interp0_pop_lane1: INTERP0_POP_LANE1,
    pub interp0_pop_full: INTERP0_POP_FULL,
    pub interp0_peek_lane0: INTERP0_PEEK_LANE0,
    pub interp0_peek_lane1: INTERP0_PEEK_LANE1,
    pub interp0_peek_full: INTERP0_PEEK_FULL,
    pub interp0_ctrl_lane0: INTERP0_CTRL_LANE0,
    pub interp0_ctrl_lane1: INTERP0_CTRL_LANE1,
    pub interp0_accum0_add: INTERP0_ACCUM0_ADD,
    pub interp0_accum1_add: INTERP0_ACCUM1_ADD,
    pub interp0_base_1and0: INTERP0_BASE_1AND0,
    pub interp1_accum0: INTERP1_ACCUM0,
    pub interp1_accum1: INTERP1_ACCUM1,
    pub interp1_base0: INTERP1_BASE0,
    pub interp1_base1: INTERP1_BASE1,
    pub interp1_base2: INTERP1_BASE2,
    pub interp1_pop_lane0: INTERP1_POP_LANE0,
    pub interp1_pop_lane1: INTERP1_POP_LANE1,
    pub interp1_pop_full: INTERP1_POP_FULL,
    pub interp1_peek_lane0: INTERP1_PEEK_LANE0,
    pub interp1_peek_lane1: INTERP1_PEEK_LANE1,
    pub interp1_peek_full: INTERP1_PEEK_FULL,
    pub interp1_ctrl_lane0: INTERP1_CTRL_LANE0,
    pub interp1_ctrl_lane1: INTERP1_CTRL_LANE1,
    pub interp1_accum0_add: INTERP1_ACCUM0_ADD,
    pub interp1_accum1_add: INTERP1_ACCUM1_ADD,
    pub interp1_base_1and0: INTERP1_BASE_1AND0,
    pub spinlock0: SPINLOCK0,
    pub spinlock1: SPINLOCK1,
    pub spinlock2: SPINLOCK2,
    pub spinlock3: SPINLOCK3,
    pub spinlock4: SPINLOCK4,
    pub spinlock5: SPINLOCK5,
    pub spinlock6: SPINLOCK6,
    pub spinlock7: SPINLOCK7,
    pub spinlock8: SPINLOCK8,
    pub spinlock9: SPINLOCK9,
    pub spinlock10: SPINLOCK10,
    pub spinlock11: SPINLOCK11,
    pub spinlock12: SPINLOCK12,
    pub spinlock13: SPINLOCK13,
    pub spinlock14: SPINLOCK14,
    pub spinlock15: SPINLOCK15,
    pub spinlock16: SPINLOCK16,
    pub spinlock17: SPINLOCK17,
    pub spinlock18: SPINLOCK18,
    pub spinlock19: SPINLOCK19,
    pub spinlock20: SPINLOCK20,
    pub spinlock21: SPINLOCK21,
    pub spinlock22: SPINLOCK22,
    pub spinlock23: SPINLOCK23,
    pub spinlock24: SPINLOCK24,
    pub spinlock25: SPINLOCK25,
    pub spinlock26: SPINLOCK26,
    pub spinlock27: SPINLOCK27,
    pub spinlock28: SPINLOCK28,
    pub spinlock29: SPINLOCK29,
    pub spinlock30: SPINLOCK30,
    pub spinlock31: SPINLOCK31,
    // some fields omitted
}

Register block

Fields
cpuid: CPUID

0x00 - Processor core identifier\n Value is 0 when read from processor core 0, and 1 when read from processor core 1.

gpio_in: GPIO_IN

0x04 - Input value for GPIO pins

gpio_hi_in: GPIO_HI_IN

0x08 - Input value for QSPI pins

gpio_out: GPIO_OUT

0x10 - GPIO output value

gpio_out_set: GPIO_OUT_SET

0x14 - GPIO output value set

gpio_out_clr: GPIO_OUT_CLR

0x18 - GPIO output value clear

gpio_out_xor: GPIO_OUT_XOR

0x1c - GPIO output value XOR

gpio_oe: GPIO_OE

0x20 - GPIO output enable

gpio_oe_set: GPIO_OE_SET

0x24 - GPIO output enable set

gpio_oe_clr: GPIO_OE_CLR

0x28 - GPIO output enable clear

gpio_oe_xor: GPIO_OE_XOR

0x2c - GPIO output enable XOR

gpio_hi_out: GPIO_HI_OUT

0x30 - QSPI output value

gpio_hi_out_set: GPIO_HI_OUT_SET

0x34 - QSPI output value set

gpio_hi_out_clr: GPIO_HI_OUT_CLR

0x38 - QSPI output value clear

gpio_hi_out_xor: GPIO_HI_OUT_XOR

0x3c - QSPI output value XOR

gpio_hi_oe: GPIO_HI_OE

0x40 - QSPI output enable

gpio_hi_oe_set: GPIO_HI_OE_SET

0x44 - QSPI output enable set

gpio_hi_oe_clr: GPIO_HI_OE_CLR

0x48 - QSPI output enable clear

gpio_hi_oe_xor: GPIO_HI_OE_XOR

0x4c - QSPI output enable XOR

fifo_st: FIFO_ST

0x50 - Status register for inter-core FIFOs (mailboxes).\n There is one FIFO in the core 0 -> core 1 direction, and one core 1 -> core 0. Both are 32 bits wide and 8 words deep.\n Core 0 can see the read side of the 1->0 FIFO (RX), and the write side of 0->1 FIFO (TX).\n Core 1 can see the read side of the 0->1 FIFO (RX), and the write side of 1->0 FIFO (TX).\n The SIO IRQ for each core is the logical OR of the VLD, WOF and ROE fields of its FIFO_ST register.

fifo_wr: FIFO_WR

0x54 - Write access to this core’s TX FIFO

fifo_rd: FIFO_RD

0x58 - Read access to this core’s RX FIFO

spinlock_st: SPINLOCK_ST

0x5c - Spinlock state\n A bitmap containing the state of all 32 spinlocks (1=locked).\n Mainly intended for debugging.

div_udividend: DIV_UDIVIDEND

0x60 - Divider unsigned dividend\n Write to the DIVIDEND operand of the divider, i.e. the p in p / q.\n Any operand write starts a new calculation. The results appear in QUOTIENT, REMAINDER.\n UDIVIDEND/SDIVIDEND are aliases of the same internal register. The U alias starts an\n unsigned calculation, and the S alias starts a signed calculation.

div_udivisor: DIV_UDIVISOR

0x64 - Divider unsigned divisor\n Write to the DIVISOR operand of the divider, i.e. the q in p / q.\n Any operand write starts a new calculation. The results appear in QUOTIENT, REMAINDER.\n UDIVIDEND/SDIVIDEND are aliases of the same internal register. The U alias starts an\n unsigned calculation, and the S alias starts a signed calculation.

div_sdividend: DIV_SDIVIDEND

0x68 - Divider signed dividend\n The same as UDIVIDEND, but starts a signed calculation, rather than unsigned.

div_sdivisor: DIV_SDIVISOR

0x6c - Divider signed divisor\n The same as UDIVISOR, but starts a signed calculation, rather than unsigned.

div_quotient: DIV_QUOTIENT

0x70 - Divider result quotient\n The result of DIVIDEND / DIVISOR (division). Contents undefined while CSR_READY is low.\n For signed calculations, QUOTIENT is negative when the signs of DIVIDEND and DIVISOR differ.\n This register can be written to directly, for context save/restore purposes. This halts any\n in-progress calculation and sets the CSR_READY and CSR_DIRTY flags.\n Reading from QUOTIENT clears the CSR_DIRTY flag, so should read results in the order\n REMAINDER, QUOTIENT if CSR_DIRTY is used.

div_remainder: DIV_REMAINDER

0x74 - Divider result remainder\n The result of DIVIDEND % DIVISOR (modulo). Contents undefined while CSR_READY is low.\n For signed calculations, REMAINDER is negative only when DIVIDEND is negative.\n This register can be written to directly, for context save/restore purposes. This halts any\n in-progress calculation and sets the CSR_READY and CSR_DIRTY flags.

div_csr: DIV_CSR

0x78 - Control and status register for divider.

interp0_accum0: INTERP0_ACCUM0

0x80 - Read/write access to accumulator 0

interp0_accum1: INTERP0_ACCUM1

0x84 - Read/write access to accumulator 1

interp0_base0: INTERP0_BASE0

0x88 - Read/write access to BASE0 register.

interp0_base1: INTERP0_BASE1

0x8c - Read/write access to BASE1 register.

interp0_base2: INTERP0_BASE2

0x90 - Read/write access to BASE2 register.

interp0_pop_lane0: INTERP0_POP_LANE0

0x94 - Read LANE0 result, and simultaneously write lane results to both accumulators (POP).

interp0_pop_lane1: INTERP0_POP_LANE1

0x98 - Read LANE1 result, and simultaneously write lane results to both accumulators (POP).

interp0_pop_full: INTERP0_POP_FULL

0x9c - Read FULL result, and simultaneously write lane results to both accumulators (POP).

interp0_peek_lane0: INTERP0_PEEK_LANE0

0xa0 - Read LANE0 result, without altering any internal state (PEEK).

interp0_peek_lane1: INTERP0_PEEK_LANE1

0xa4 - Read LANE1 result, without altering any internal state (PEEK).

interp0_peek_full: INTERP0_PEEK_FULL

0xa8 - Read FULL result, without altering any internal state (PEEK).

interp0_ctrl_lane0: INTERP0_CTRL_LANE0

0xac - Control register for lane 0

interp0_ctrl_lane1: INTERP0_CTRL_LANE1

0xb0 - Control register for lane 1

interp0_accum0_add: INTERP0_ACCUM0_ADD

0xb4 - Values written here are atomically added to ACCUM0\n Reading yields lane 0’s raw shift and mask value (BASE0 not added).

interp0_accum1_add: INTERP0_ACCUM1_ADD

0xb8 - Values written here are atomically added to ACCUM1\n Reading yields lane 1’s raw shift and mask value (BASE1 not added).

interp0_base_1and0: INTERP0_BASE_1AND0

0xbc - On write, the lower 16 bits go to BASE0, upper bits to BASE1 simultaneously.\n Each half is sign-extended to 32 bits if that lane’s SIGNED flag is set.

interp1_accum0: INTERP1_ACCUM0

0xc0 - Read/write access to accumulator 0

interp1_accum1: INTERP1_ACCUM1

0xc4 - Read/write access to accumulator 1

interp1_base0: INTERP1_BASE0

0xc8 - Read/write access to BASE0 register.

interp1_base1: INTERP1_BASE1

0xcc - Read/write access to BASE1 register.

interp1_base2: INTERP1_BASE2

0xd0 - Read/write access to BASE2 register.

interp1_pop_lane0: INTERP1_POP_LANE0

0xd4 - Read LANE0 result, and simultaneously write lane results to both accumulators (POP).

interp1_pop_lane1: INTERP1_POP_LANE1

0xd8 - Read LANE1 result, and simultaneously write lane results to both accumulators (POP).

interp1_pop_full: INTERP1_POP_FULL

0xdc - Read FULL result, and simultaneously write lane results to both accumulators (POP).

interp1_peek_lane0: INTERP1_PEEK_LANE0

0xe0 - Read LANE0 result, without altering any internal state (PEEK).

interp1_peek_lane1: INTERP1_PEEK_LANE1

0xe4 - Read LANE1 result, without altering any internal state (PEEK).

interp1_peek_full: INTERP1_PEEK_FULL

0xe8 - Read FULL result, without altering any internal state (PEEK).

interp1_ctrl_lane0: INTERP1_CTRL_LANE0

0xec - Control register for lane 0

interp1_ctrl_lane1: INTERP1_CTRL_LANE1

0xf0 - Control register for lane 1

interp1_accum0_add: INTERP1_ACCUM0_ADD

0xf4 - Values written here are atomically added to ACCUM0\n Reading yields lane 0’s raw shift and mask value (BASE0 not added).

interp1_accum1_add: INTERP1_ACCUM1_ADD

0xf8 - Values written here are atomically added to ACCUM1\n Reading yields lane 1’s raw shift and mask value (BASE1 not added).

interp1_base_1and0: INTERP1_BASE_1AND0

0xfc - On write, the lower 16 bits go to BASE0, upper bits to BASE1 simultaneously.\n Each half is sign-extended to 32 bits if that lane’s SIGNED flag is set.

spinlock0: SPINLOCK0

0x100 - Reading from a spinlock address will:\n - Return 0 if lock is already locked\n - Otherwise return nonzero, and simultaneously claim the lock\n\n Writing (any value) releases the lock.\n If core 0 and core 1 attempt to claim the same lock simultaneously, core 0 wins.\n The value returned on success is 0x1 << lock number.

spinlock1: SPINLOCK1

0x104 - Reading from a spinlock address will:\n - Return 0 if lock is already locked\n - Otherwise return nonzero, and simultaneously claim the lock\n\n Writing (any value) releases the lock.\n If core 0 and core 1 attempt to claim the same lock simultaneously, core 0 wins.\n The value returned on success is 0x1 << lock number.

spinlock2: SPINLOCK2

0x108 - Reading from a spinlock address will:\n - Return 0 if lock is already locked\n - Otherwise return nonzero, and simultaneously claim the lock\n\n Writing (any value) releases the lock.\n If core 0 and core 1 attempt to claim the same lock simultaneously, core 0 wins.\n The value returned on success is 0x1 << lock number.

spinlock3: SPINLOCK3

0x10c - Reading from a spinlock address will:\n - Return 0 if lock is already locked\n - Otherwise return nonzero, and simultaneously claim the lock\n\n Writing (any value) releases the lock.\n If core 0 and core 1 attempt to claim the same lock simultaneously, core 0 wins.\n The value returned on success is 0x1 << lock number.

spinlock4: SPINLOCK4

0x110 - Reading from a spinlock address will:\n - Return 0 if lock is already locked\n - Otherwise return nonzero, and simultaneously claim the lock\n\n Writing (any value) releases the lock.\n If core 0 and core 1 attempt to claim the same lock simultaneously, core 0 wins.\n The value returned on success is 0x1 << lock number.

spinlock5: SPINLOCK5

0x114 - Reading from a spinlock address will:\n - Return 0 if lock is already locked\n - Otherwise return nonzero, and simultaneously claim the lock\n\n Writing (any value) releases the lock.\n If core 0 and core 1 attempt to claim the same lock simultaneously, core 0 wins.\n The value returned on success is 0x1 << lock number.

spinlock6: SPINLOCK6

0x118 - Reading from a spinlock address will:\n - Return 0 if lock is already locked\n - Otherwise return nonzero, and simultaneously claim the lock\n\n Writing (any value) releases the lock.\n If core 0 and core 1 attempt to claim the same lock simultaneously, core 0 wins.\n The value returned on success is 0x1 << lock number.

spinlock7: SPINLOCK7

0x11c - Reading from a spinlock address will:\n - Return 0 if lock is already locked\n - Otherwise return nonzero, and simultaneously claim the lock\n\n Writing (any value) releases the lock.\n If core 0 and core 1 attempt to claim the same lock simultaneously, core 0 wins.\n The value returned on success is 0x1 << lock number.

spinlock8: SPINLOCK8

0x120 - Reading from a spinlock address will:\n - Return 0 if lock is already locked\n - Otherwise return nonzero, and simultaneously claim the lock\n\n Writing (any value) releases the lock.\n If core 0 and core 1 attempt to claim the same lock simultaneously, core 0 wins.\n The value returned on success is 0x1 << lock number.

spinlock9: SPINLOCK9

0x124 - Reading from a spinlock address will:\n - Return 0 if lock is already locked\n - Otherwise return nonzero, and simultaneously claim the lock\n\n Writing (any value) releases the lock.\n If core 0 and core 1 attempt to claim the same lock simultaneously, core 0 wins.\n The value returned on success is 0x1 << lock number.

spinlock10: SPINLOCK10

0x128 - Reading from a spinlock address will:\n - Return 0 if lock is already locked\n - Otherwise return nonzero, and simultaneously claim the lock\n\n Writing (any value) releases the lock.\n If core 0 and core 1 attempt to claim the same lock simultaneously, core 0 wins.\n The value returned on success is 0x1 << lock number.

spinlock11: SPINLOCK11

0x12c - Reading from a spinlock address will:\n - Return 0 if lock is already locked\n - Otherwise return nonzero, and simultaneously claim the lock\n\n Writing (any value) releases the lock.\n If core 0 and core 1 attempt to claim the same lock simultaneously, core 0 wins.\n The value returned on success is 0x1 << lock number.

spinlock12: SPINLOCK12

0x130 - Reading from a spinlock address will:\n - Return 0 if lock is already locked\n - Otherwise return nonzero, and simultaneously claim the lock\n\n Writing (any value) releases the lock.\n If core 0 and core 1 attempt to claim the same lock simultaneously, core 0 wins.\n The value returned on success is 0x1 << lock number.

spinlock13: SPINLOCK13

0x134 - Reading from a spinlock address will:\n - Return 0 if lock is already locked\n - Otherwise return nonzero, and simultaneously claim the lock\n\n Writing (any value) releases the lock.\n If core 0 and core 1 attempt to claim the same lock simultaneously, core 0 wins.\n The value returned on success is 0x1 << lock number.

spinlock14: SPINLOCK14

0x138 - Reading from a spinlock address will:\n - Return 0 if lock is already locked\n - Otherwise return nonzero, and simultaneously claim the lock\n\n Writing (any value) releases the lock.\n If core 0 and core 1 attempt to claim the same lock simultaneously, core 0 wins.\n The value returned on success is 0x1 << lock number.

spinlock15: SPINLOCK15

0x13c - Reading from a spinlock address will:\n - Return 0 if lock is already locked\n - Otherwise return nonzero, and simultaneously claim the lock\n\n Writing (any value) releases the lock.\n If core 0 and core 1 attempt to claim the same lock simultaneously, core 0 wins.\n The value returned on success is 0x1 << lock number.

spinlock16: SPINLOCK16

0x140 - Reading from a spinlock address will:\n - Return 0 if lock is already locked\n - Otherwise return nonzero, and simultaneously claim the lock\n\n Writing (any value) releases the lock.\n If core 0 and core 1 attempt to claim the same lock simultaneously, core 0 wins.\n The value returned on success is 0x1 << lock number.

spinlock17: SPINLOCK17

0x144 - Reading from a spinlock address will:\n - Return 0 if lock is already locked\n - Otherwise return nonzero, and simultaneously claim the lock\n\n Writing (any value) releases the lock.\n If core 0 and core 1 attempt to claim the same lock simultaneously, core 0 wins.\n The value returned on success is 0x1 << lock number.

spinlock18: SPINLOCK18

0x148 - Reading from a spinlock address will:\n - Return 0 if lock is already locked\n - Otherwise return nonzero, and simultaneously claim the lock\n\n Writing (any value) releases the lock.\n If core 0 and core 1 attempt to claim the same lock simultaneously, core 0 wins.\n The value returned on success is 0x1 << lock number.

spinlock19: SPINLOCK19

0x14c - Reading from a spinlock address will:\n - Return 0 if lock is already locked\n - Otherwise return nonzero, and simultaneously claim the lock\n\n Writing (any value) releases the lock.\n If core 0 and core 1 attempt to claim the same lock simultaneously, core 0 wins.\n The value returned on success is 0x1 << lock number.

spinlock20: SPINLOCK20

0x150 - Reading from a spinlock address will:\n - Return 0 if lock is already locked\n - Otherwise return nonzero, and simultaneously claim the lock\n\n Writing (any value) releases the lock.\n If core 0 and core 1 attempt to claim the same lock simultaneously, core 0 wins.\n The value returned on success is 0x1 << lock number.

spinlock21: SPINLOCK21

0x154 - Reading from a spinlock address will:\n - Return 0 if lock is already locked\n - Otherwise return nonzero, and simultaneously claim the lock\n\n Writing (any value) releases the lock.\n If core 0 and core 1 attempt to claim the same lock simultaneously, core 0 wins.\n The value returned on success is 0x1 << lock number.

spinlock22: SPINLOCK22

0x158 - Reading from a spinlock address will:\n - Return 0 if lock is already locked\n - Otherwise return nonzero, and simultaneously claim the lock\n\n Writing (any value) releases the lock.\n If core 0 and core 1 attempt to claim the same lock simultaneously, core 0 wins.\n The value returned on success is 0x1 << lock number.

spinlock23: SPINLOCK23

0x15c - Reading from a spinlock address will:\n - Return 0 if lock is already locked\n - Otherwise return nonzero, and simultaneously claim the lock\n\n Writing (any value) releases the lock.\n If core 0 and core 1 attempt to claim the same lock simultaneously, core 0 wins.\n The value returned on success is 0x1 << lock number.

spinlock24: SPINLOCK24

0x160 - Reading from a spinlock address will:\n - Return 0 if lock is already locked\n - Otherwise return nonzero, and simultaneously claim the lock\n\n Writing (any value) releases the lock.\n If core 0 and core 1 attempt to claim the same lock simultaneously, core 0 wins.\n The value returned on success is 0x1 << lock number.

spinlock25: SPINLOCK25

0x164 - Reading from a spinlock address will:\n - Return 0 if lock is already locked\n - Otherwise return nonzero, and simultaneously claim the lock\n\n Writing (any value) releases the lock.\n If core 0 and core 1 attempt to claim the same lock simultaneously, core 0 wins.\n The value returned on success is 0x1 << lock number.

spinlock26: SPINLOCK26

0x168 - Reading from a spinlock address will:\n - Return 0 if lock is already locked\n - Otherwise return nonzero, and simultaneously claim the lock\n\n Writing (any value) releases the lock.\n If core 0 and core 1 attempt to claim the same lock simultaneously, core 0 wins.\n The value returned on success is 0x1 << lock number.

spinlock27: SPINLOCK27

0x16c - Reading from a spinlock address will:\n - Return 0 if lock is already locked\n - Otherwise return nonzero, and simultaneously claim the lock\n\n Writing (any value) releases the lock.\n If core 0 and core 1 attempt to claim the same lock simultaneously, core 0 wins.\n The value returned on success is 0x1 << lock number.

spinlock28: SPINLOCK28

0x170 - Reading from a spinlock address will:\n - Return 0 if lock is already locked\n - Otherwise return nonzero, and simultaneously claim the lock\n\n Writing (any value) releases the lock.\n If core 0 and core 1 attempt to claim the same lock simultaneously, core 0 wins.\n The value returned on success is 0x1 << lock number.

spinlock29: SPINLOCK29

0x174 - Reading from a spinlock address will:\n - Return 0 if lock is already locked\n - Otherwise return nonzero, and simultaneously claim the lock\n\n Writing (any value) releases the lock.\n If core 0 and core 1 attempt to claim the same lock simultaneously, core 0 wins.\n The value returned on success is 0x1 << lock number.

spinlock30: SPINLOCK30

0x178 - Reading from a spinlock address will:\n - Return 0 if lock is already locked\n - Otherwise return nonzero, and simultaneously claim the lock\n\n Writing (any value) releases the lock.\n If core 0 and core 1 attempt to claim the same lock simultaneously, core 0 wins.\n The value returned on success is 0x1 << lock number.

spinlock31: SPINLOCK31

0x17c - Reading from a spinlock address will:\n - Return 0 if lock is already locked\n - Otherwise return nonzero, and simultaneously claim the lock\n\n Writing (any value) releases the lock.\n If core 0 and core 1 attempt to claim the same lock simultaneously, core 0 wins.\n The value returned on success is 0x1 << lock number.