Type Definition rp2040::spi0::sspimsc::W

type W = W<u32, SSPIMSC>;

Writer for register SSPIMSC

Implementations

impl W

pub fn txim(&mut self) -> TXIM_W<'_>

Bit 3 - Transmit FIFO interrupt mask: 0 Transmit FIFO half empty or less condition interrupt is masked. 1 Transmit FIFO half empty or less condition interrupt is not masked.

pub fn rxim(&mut self) -> RXIM_W<'_>

Bit 2 - Receive FIFO interrupt mask: 0 Receive FIFO half full or less condition interrupt is masked. 1 Receive FIFO half full or less condition interrupt is not masked.

pub fn rtim(&mut self) -> RTIM_W<'_>

Bit 1 - Receive timeout interrupt mask: 0 Receive FIFO not empty and no read prior to timeout period interrupt is masked. 1 Receive FIFO not empty and no read prior to timeout period interrupt is not masked.

pub fn rorim(&mut self) -> RORIM_W<'_>

Bit 0 - Receive overrun interrupt mask: 0 Receive FIFO written to while full condition interrupt is masked. 1 Receive FIFO written to while full condition interrupt is not masked.