Type Definition rp2040::spi0::sspmis::R

type R = R<u32, SSPMIS>;

Reader of register SSPMIS

Implementations

impl R

pub fn txmis(&self) -> TXMIS_R

Bit 3 - Gives the transmit FIFO masked interrupt state, after masking, of the SSPTXINTR interrupt

pub fn rxmis(&self) -> RXMIS_R

Bit 2 - Gives the receive FIFO masked interrupt state, after masking, of the SSPRXINTR interrupt

pub fn rtmis(&self) -> RTMIS_R

Bit 1 - Gives the receive timeout masked interrupt state, after masking, of the SSPRTINTR interrupt

pub fn rormis(&self) -> RORMIS_R

Bit 0 - Gives the receive over run masked interrupt status, after masking, of the SSPRORINTR interrupt