Type Definition rp2040::syscfg::dbgforce::R

type R = R<u32, DBGFORCE>;

Reader of register DBGFORCE

Implementations

impl R

pub fn proc1_attach(&self) -> PROC1_ATTACH_R

Bit 7 - Attach processor 1 debug port to syscfg controls, and disconnect it from external SWD pads.

pub fn proc1_swclk(&self) -> PROC1_SWCLK_R

Bit 6 - Directly drive processor 1 SWCLK, if PROC1_ATTACH is set

pub fn proc1_swdi(&self) -> PROC1_SWDI_R

Bit 5 - Directly drive processor 1 SWDIO input, if PROC1_ATTACH is set

pub fn proc1_swdo(&self) -> PROC1_SWDO_R

Bit 4 - Observe the value of processor 1 SWDIO output.

pub fn proc0_attach(&self) -> PROC0_ATTACH_R

Bit 3 - Attach processor 0 debug port to syscfg controls, and disconnect it from external SWD pads.

pub fn proc0_swclk(&self) -> PROC0_SWCLK_R

Bit 2 - Directly drive processor 0 SWCLK, if PROC0_ATTACH is set

pub fn proc0_swdi(&self) -> PROC0_SWDI_R

Bit 1 - Directly drive processor 0 SWDIO input, if PROC0_ATTACH is set

pub fn proc0_swdo(&self) -> PROC0_SWDO_R

Bit 0 - Observe the value of processor 0 SWDIO output.