Type Definition rp2040::syscfg::proc_config::R

type R = R<u32, PROC_CONFIG>;

Reader of register PROC_CONFIG

Implementations

impl R

pub fn proc1_dap_instid(&self) -> PROC1_DAP_INSTID_R

Bits 28:31 - Configure proc1 DAP instance ID.\n Recommend that this is NOT changed until you require debug access in multi-chip environment\n WARNING: do not set to 15 as this is reserved for RescueDP

pub fn proc0_dap_instid(&self) -> PROC0_DAP_INSTID_R

Bits 24:27 - Configure proc0 DAP instance ID.\n Recommend that this is NOT changed until you require debug access in multi-chip environment\n WARNING: do not set to 15 as this is reserved for RescueDP

pub fn proc1_halted(&self) -> PROC1_HALTED_R

Bit 1 - Indication that proc1 has halted

pub fn proc0_halted(&self) -> PROC0_HALTED_R

Bit 0 - Indication that proc0 has halted