Module rp2040::usbctrl_regs::buff_status

Buffer status register. A bit set here indicates that a buffer has completed on the endpoint (if the buffer interrupt is enabled). It is possible for 2 buffers to be completed, so clearing the buffer status bit may instantly re set it on the next clock cycle.

Type Definitions

EP0_IN_R
Reader of field EP0_IN

EP0_OUT_R
Reader of field EP0_OUT

EP1_IN_R
Reader of field EP1_IN

EP1_OUT_R
Reader of field EP1_OUT

EP2_IN_R
Reader of field EP2_IN

EP2_OUT_R
Reader of field EP2_OUT

EP3_IN_R
Reader of field EP3_IN

EP3_OUT_R
Reader of field EP3_OUT

EP4_IN_R
Reader of field EP4_IN

EP4_OUT_R
Reader of field EP4_OUT

EP5_IN_R
Reader of field EP5_IN

EP5_OUT_R
Reader of field EP5_OUT

EP6_IN_R
Reader of field EP6_IN

EP6_OUT_R
Reader of field EP6_OUT

EP7_IN_R
Reader of field EP7_IN

EP7_OUT_R
Reader of field EP7_OUT

EP8_IN_R
Reader of field EP8_IN

EP8_OUT_R
Reader of field EP8_OUT

EP9_IN_R
Reader of field EP9_IN

EP9_OUT_R
Reader of field EP9_OUT

EP10_IN_R
Reader of field EP10_IN

EP10_OUT_R
Reader of field EP10_OUT

EP11_IN_R
Reader of field EP11_IN

EP11_OUT_R
Reader of field EP11_OUT

EP12_IN_R
Reader of field EP12_IN

EP12_OUT_R
Reader of field EP12_OUT

EP13_IN_R
Reader of field EP13_IN

EP13_OUT_R
Reader of field EP13_OUT

EP14_IN_R
Reader of field EP14_IN

EP14_OUT_R
Reader of field EP14_OUT

EP15_IN_R
Reader of field EP15_IN

EP15_OUT_R
Reader of field EP15_OUT

R
Reader of register BUFF_STATUS