Module rp2040::usbctrl_regs::ep_abort_done

Device only: Used in conjunction with EP_ABORT. Set once an endpoint is idle so the programmer knows it is safe to modify the buffer control register.

Structs

EP0_IN_W
Write proxy for field EP0_IN

EP0_OUT_W
Write proxy for field EP0_OUT

EP1_IN_W
Write proxy for field EP1_IN

EP1_OUT_W
Write proxy for field EP1_OUT

EP2_IN_W
Write proxy for field EP2_IN

EP2_OUT_W
Write proxy for field EP2_OUT

EP3_IN_W
Write proxy for field EP3_IN

EP3_OUT_W
Write proxy for field EP3_OUT

EP4_IN_W
Write proxy for field EP4_IN

EP4_OUT_W
Write proxy for field EP4_OUT

EP5_IN_W
Write proxy for field EP5_IN

EP5_OUT_W
Write proxy for field EP5_OUT

EP6_IN_W
Write proxy for field EP6_IN

EP6_OUT_W
Write proxy for field EP6_OUT

EP7_IN_W
Write proxy for field EP7_IN

EP7_OUT_W
Write proxy for field EP7_OUT

EP8_IN_W
Write proxy for field EP8_IN

EP8_OUT_W
Write proxy for field EP8_OUT

EP9_IN_W
Write proxy for field EP9_IN

EP9_OUT_W
Write proxy for field EP9_OUT

EP10_IN_W
Write proxy for field EP10_IN

EP10_OUT_W
Write proxy for field EP10_OUT

EP11_IN_W
Write proxy for field EP11_IN

EP11_OUT_W
Write proxy for field EP11_OUT

EP12_IN_W
Write proxy for field EP12_IN

EP12_OUT_W
Write proxy for field EP12_OUT

EP13_IN_W
Write proxy for field EP13_IN

EP13_OUT_W
Write proxy for field EP13_OUT

EP14_IN_W
Write proxy for field EP14_IN

EP14_OUT_W
Write proxy for field EP14_OUT

EP15_IN_W
Write proxy for field EP15_IN

EP15_OUT_W
Write proxy for field EP15_OUT

Type Definitions

EP0_IN_R
Reader of field EP0_IN

EP0_OUT_R
Reader of field EP0_OUT

EP1_IN_R
Reader of field EP1_IN

EP1_OUT_R
Reader of field EP1_OUT

EP2_IN_R
Reader of field EP2_IN

EP2_OUT_R
Reader of field EP2_OUT

EP3_IN_R
Reader of field EP3_IN

EP3_OUT_R
Reader of field EP3_OUT

EP4_IN_R
Reader of field EP4_IN

EP4_OUT_R
Reader of field EP4_OUT

EP5_IN_R
Reader of field EP5_IN

EP5_OUT_R
Reader of field EP5_OUT

EP6_IN_R
Reader of field EP6_IN

EP6_OUT_R
Reader of field EP6_OUT

EP7_IN_R
Reader of field EP7_IN

EP7_OUT_R
Reader of field EP7_OUT

EP8_IN_R
Reader of field EP8_IN

EP8_OUT_R
Reader of field EP8_OUT

EP9_IN_R
Reader of field EP9_IN

EP9_OUT_R
Reader of field EP9_OUT

EP10_IN_R
Reader of field EP10_IN

EP10_OUT_R
Reader of field EP10_OUT

EP11_IN_R
Reader of field EP11_IN

EP11_OUT_R
Reader of field EP11_OUT

EP12_IN_R
Reader of field EP12_IN

EP12_OUT_R
Reader of field EP12_OUT

EP13_IN_R
Reader of field EP13_IN

EP13_OUT_R
Reader of field EP13_OUT

EP14_IN_R
Reader of field EP14_IN

EP14_OUT_R
Reader of field EP14_OUT

EP15_IN_R
Reader of field EP15_IN

EP15_OUT_R
Reader of field EP15_OUT

R
Reader of register EP_ABORT_DONE

W
Writer for register EP_ABORT_DONE