Type Definition rp2040::usbctrl_regs::sie_status::W

type W = W<u32, SIE_STATUS>;

Writer for register SIE_STATUS

Implementations

impl W

pub fn data_seq_error(&mut self) -> DATA_SEQ_ERROR_W<'_>

Bit 31 - Data Sequence Error.\n\n The device can raise a sequence error in the following conditions:\n\n * A SETUP packet is received followed by a DATA1 packet (data phase should always be DATA0) * An OUT packet is received from the host but doesn’t match the data pid in the buffer control register read from DPSRAM\n\n The host can raise a data sequence error in the following conditions:\n\n * An IN packet from the device has the wrong data PID

pub fn ack_rec(&mut self) -> ACK_REC_W<'_>

Bit 30 - ACK received. Raised by both host and device.

pub fn stall_rec(&mut self) -> STALL_REC_W<'_>

Bit 29 - Host: STALL received

pub fn nak_rec(&mut self) -> NAK_REC_W<'_>

Bit 28 - Host: NAK received

pub fn rx_timeout(&mut self) -> RX_TIMEOUT_W<'_>

Bit 27 - RX timeout is raised by both the host and device if an ACK is not received in the maximum time specified by the USB spec.

pub fn rx_overflow(&mut self) -> RX_OVERFLOW_W<'_>

Bit 26 - RX overflow is raised by the Serial RX engine if the incoming data is too fast.

pub fn bit_stuff_error(&mut self) -> BIT_STUFF_ERROR_W<'_>

Bit 25 - Bit Stuff Error. Raised by the Serial RX engine.

pub fn crc_error(&mut self) -> CRC_ERROR_W<'_>

Bit 24 - CRC Error. Raised by the Serial RX engine.

pub fn bus_reset(&mut self) -> BUS_RESET_W<'_>

Bit 19 - Device: bus reset received

pub fn trans_complete(&mut self) -> TRANS_COMPLETE_W<'_>

Bit 18 - Transaction complete.\n\n Raised by device if:\n\n * An IN or OUT packet is sent with the LAST_BUFF bit set in the buffer control register\n\n Raised by host if:\n\n * A setup packet is sent when no data in or data out transaction follows * An IN packet is received and the LAST_BUFF bit is set in the buffer control register * An IN packet is received with zero length * An OUT packet is sent and the LAST_BUFF bit is set

pub fn setup_rec(&mut self) -> SETUP_REC_W<'_>

Bit 17 - Device: Setup packet received

pub fn resume(&mut self) -> RESUME_W<'_>

Bit 11 - Host: Device has initiated a remote resume. Device: host has initiated a resume.