Struct rp2040::usbctrl_regs::RegisterBlock

#[repr(C)]pub struct RegisterBlock {
    pub addr_endp: ADDR_ENDP,
    pub addr_endp1: ADDR_ENDP1,
    pub addr_endp2: ADDR_ENDP2,
    pub addr_endp3: ADDR_ENDP3,
    pub addr_endp4: ADDR_ENDP4,
    pub addr_endp5: ADDR_ENDP5,
    pub addr_endp6: ADDR_ENDP6,
    pub addr_endp7: ADDR_ENDP7,
    pub addr_endp8: ADDR_ENDP8,
    pub addr_endp9: ADDR_ENDP9,
    pub addr_endp10: ADDR_ENDP10,
    pub addr_endp11: ADDR_ENDP11,
    pub addr_endp12: ADDR_ENDP12,
    pub addr_endp13: ADDR_ENDP13,
    pub addr_endp14: ADDR_ENDP14,
    pub addr_endp15: ADDR_ENDP15,
    pub main_ctrl: MAIN_CTRL,
    pub sof_wr: SOF_WR,
    pub sof_rd: SOF_RD,
    pub sie_ctrl: SIE_CTRL,
    pub sie_status: SIE_STATUS,
    pub int_ep_ctrl: INT_EP_CTRL,
    pub buff_status: BUFF_STATUS,
    pub buff_cpu_should_handle: BUFF_CPU_SHOULD_HANDLE,
    pub ep_abort: EP_ABORT,
    pub ep_abort_done: EP_ABORT_DONE,
    pub ep_stall_arm: EP_STALL_ARM,
    pub nak_poll: NAK_POLL,
    pub ep_status_stall_nak: EP_STATUS_STALL_NAK,
    pub usb_muxing: USB_MUXING,
    pub usb_pwr: USB_PWR,
    pub usbphy_direct: USBPHY_DIRECT,
    pub usbphy_direct_override: USBPHY_DIRECT_OVERRIDE,
    pub usbphy_trim: USBPHY_TRIM,
    pub intr: INTR,
    pub inte: INTE,
    pub intf: INTF,
    pub ints: INTS,
    // some fields omitted
}

Register block

Fields
addr_endp: ADDR_ENDP

0x00 - Device address and endpoint control

addr_endp1: ADDR_ENDP1

0x04 - Interrupt endpoint 1. Only valid for HOST mode.

addr_endp2: ADDR_ENDP2

0x08 - Interrupt endpoint 2. Only valid for HOST mode.

addr_endp3: ADDR_ENDP3

0x0c - Interrupt endpoint 3. Only valid for HOST mode.

addr_endp4: ADDR_ENDP4

0x10 - Interrupt endpoint 4. Only valid for HOST mode.

addr_endp5: ADDR_ENDP5

0x14 - Interrupt endpoint 5. Only valid for HOST mode.

addr_endp6: ADDR_ENDP6

0x18 - Interrupt endpoint 6. Only valid for HOST mode.

addr_endp7: ADDR_ENDP7

0x1c - Interrupt endpoint 7. Only valid for HOST mode.

addr_endp8: ADDR_ENDP8

0x20 - Interrupt endpoint 8. Only valid for HOST mode.

addr_endp9: ADDR_ENDP9

0x24 - Interrupt endpoint 9. Only valid for HOST mode.

addr_endp10: ADDR_ENDP10

0x28 - Interrupt endpoint 10. Only valid for HOST mode.

addr_endp11: ADDR_ENDP11

0x2c - Interrupt endpoint 11. Only valid for HOST mode.

addr_endp12: ADDR_ENDP12

0x30 - Interrupt endpoint 12. Only valid for HOST mode.

addr_endp13: ADDR_ENDP13

0x34 - Interrupt endpoint 13. Only valid for HOST mode.

addr_endp14: ADDR_ENDP14

0x38 - Interrupt endpoint 14. Only valid for HOST mode.

addr_endp15: ADDR_ENDP15

0x3c - Interrupt endpoint 15. Only valid for HOST mode.

main_ctrl: MAIN_CTRL

0x40 - Main control register

sof_wr: SOF_WR

0x44 - Set the SOF (Start of Frame) frame number in the host controller. The SOF packet is sent every 1ms and the host will increment the frame number by 1 each time.

sof_rd: SOF_RD

0x48 - Read the last SOF (Start of Frame) frame number seen. In device mode the last SOF received from the host. In host mode the last SOF sent by the host.

sie_ctrl: SIE_CTRL

0x4c - SIE control register

sie_status: SIE_STATUS

0x50 - SIE status register

int_ep_ctrl: INT_EP_CTRL

0x54 - interrupt endpoint control register

buff_status: BUFF_STATUS

0x58 - Buffer status register. A bit set here indicates that a buffer has completed on the endpoint (if the buffer interrupt is enabled). It is possible for 2 buffers to be completed, so clearing the buffer status bit may instantly re set it on the next clock cycle.

buff_cpu_should_handle: BUFF_CPU_SHOULD_HANDLE

0x5c - Which of the double buffers should be handled. Only valid if using an interrupt per buffer (i.e. not per 2 buffers). Not valid for host interrupt endpoint polling because they are only single buffered.

ep_abort: EP_ABORT

0x60 - Device only: Can be set to ignore the buffer control register for this endpoint in case you would like to revoke a buffer. A NAK will be sent for every access to the endpoint until this bit is cleared. A corresponding bit in EP_ABORT_DONE is set when it is safe to modify the buffer control register.

ep_abort_done: EP_ABORT_DONE

0x64 - Device only: Used in conjunction with EP_ABORT. Set once an endpoint is idle so the programmer knows it is safe to modify the buffer control register.

ep_stall_arm: EP_STALL_ARM

0x68 - Device: this bit must be set in conjunction with the STALL bit in the buffer control register to send a STALL on EP0. The device controller clears these bits when a SETUP packet is received because the USB spec requires that a STALL condition is cleared when a SETUP packet is received.

nak_poll: NAK_POLL

0x6c - Used by the host controller. Sets the wait time in microseconds before trying again if the device replies with a NAK.

ep_status_stall_nak: EP_STATUS_STALL_NAK

0x70 - Device: bits are set when the IRQ_ON_NAK or IRQ_ON_STALL bits are set. For EP0 this comes from SIE_CTRL. For all other endpoints it comes from the endpoint control register.

usb_muxing: USB_MUXING

0x74 - Where to connect the USB controller. Should be to_phy by default.

usb_pwr: USB_PWR

0x78 - Overrides for the power signals in the event that the VBUS signals are not hooked up to GPIO. Set the value of the override and then the override enable to switch over to the override value.

usbphy_direct: USBPHY_DIRECT

0x7c - This register allows for direct control of the USB phy. Use in conjunction with usbphy_direct_override register to enable each override bit.

usbphy_direct_override: USBPHY_DIRECT_OVERRIDE

0x80 - Override enable for each control in usbphy_direct

usbphy_trim: USBPHY_TRIM

0x84 - Used to adjust trim values of USB phy pull down resistors.

intr: INTR

0x8c - Raw Interrupts

inte: INTE

0x90 - Interrupt Enable

intf: INTF

0x94 - Interrupt Force

ints: INTS

0x98 - Interrupt status after masking & forcing