Module rp2040::usbctrl_regs::usb_pwr

Overrides for the power signals in the event that the VBUS signals are not hooked up to GPIO. Set the value of the override and then the override enable to switch over to the override value.

Structs

OVERCURR_DETECT_EN_W
Write proxy for field OVERCURR_DETECT_EN

OVERCURR_DETECT_W
Write proxy for field OVERCURR_DETECT

VBUS_DETECT_OVERRIDE_EN_W
Write proxy for field VBUS_DETECT_OVERRIDE_EN

VBUS_DETECT_W
Write proxy for field VBUS_DETECT

VBUS_EN_OVERRIDE_EN_W
Write proxy for field VBUS_EN_OVERRIDE_EN

VBUS_EN_W
Write proxy for field VBUS_EN

Type Definitions

OVERCURR_DETECT_EN_R
Reader of field OVERCURR_DETECT_EN

OVERCURR_DETECT_R
Reader of field OVERCURR_DETECT

R
Reader of register USB_PWR

VBUS_DETECT_OVERRIDE_EN_R
Reader of field VBUS_DETECT_OVERRIDE_EN

VBUS_DETECT_R
Reader of field VBUS_DETECT

VBUS_EN_OVERRIDE_EN_R
Reader of field VBUS_EN_OVERRIDE_EN

VBUS_EN_R
Reader of field VBUS_EN

W
Writer for register USB_PWR