Struct rp2040::watchdog::RegisterBlock

#[repr(C)]pub struct RegisterBlock {
    pub ctrl: CTRL,
    pub load: LOAD,
    pub reason: REASON,
    pub scratch0: SCRATCH0,
    pub scratch1: SCRATCH1,
    pub scratch2: SCRATCH2,
    pub scratch3: SCRATCH3,
    pub scratch4: SCRATCH4,
    pub scratch5: SCRATCH5,
    pub scratch6: SCRATCH6,
    pub scratch7: SCRATCH7,
    pub tick: TICK,
}

Register block

Fields
ctrl: CTRL

0x00 - Watchdog control\n The rst_wdsel register determines which subsystems are reset when the watchdog is triggered.\n The watchdog can be triggered in software.

load: LOAD

0x04 - Load the watchdog timer. The maximum setting is 0xffffff which corresponds to 0xffffff / 2 ticks before triggering a watchdog reset (see errata RP2040-E1).

reason: REASON

0x08 - Logs the reason for the last reset. Both bits are zero for the case of a hardware reset.

scratch0: SCRATCH0

0x0c - Scratch register. Information persists through soft reset of the chip.

scratch1: SCRATCH1

0x10 - Scratch register. Information persists through soft reset of the chip.

scratch2: SCRATCH2

0x14 - Scratch register. Information persists through soft reset of the chip.

scratch3: SCRATCH3

0x18 - Scratch register. Information persists through soft reset of the chip.

scratch4: SCRATCH4

0x1c - Scratch register. Information persists through soft reset of the chip.

scratch5: SCRATCH5

0x20 - Scratch register. Information persists through soft reset of the chip.

scratch6: SCRATCH6

0x24 - Scratch register. Information persists through soft reset of the chip.

scratch7: SCRATCH7

0x28 - Scratch register. Information persists through soft reset of the chip.

tick: TICK

0x2c - Controls the tick generator