Type Definition rp2040::xip_ctrl::ctrl::W

type W = W<u32, CTRL>;

Writer for register CTRL

Implementations

impl W

pub fn power_down(&mut self) -> POWER_DOWN_W<'_>

Bit 3 - When 1, the cache memories are powered down. They retain state,\n but can not be accessed. This reduces static power dissipation.\n Writing 1 to this bit forces CTRL_EN to 0, i.e. the cache cannot\n be enabled when powered down.\n Cache-as-SRAM accesses will produce a bus error response when\n the cache is powered down.

pub fn err_badwrite(&mut self) -> ERR_BADWRITE_W<'_>

Bit 1 - When 1, writes to any alias other than 0x0 (caching, allocating)\n will produce a bus fault. When 0, these writes are silently ignored.\n In either case, writes to the 0x0 alias will deallocate on tag match,\n as usual.

pub fn en(&mut self) -> EN_W<'_>

Bit 0 - When 1, enable the cache. When the cache is disabled, all XIP accesses\n will go straight to the flash, without querying the cache. When enabled,\n cacheable XIP accesses will query the cache, and the flash will\n not be accessed if the tag matches and the valid bit is set.\n\n If the cache is enabled, cache-as-SRAM accesses have no effect on the\n cache data RAM, and will produce a bus error response.