Type Definition rp2040::xip_ssi::ctrlr0::TMOD_R

type TMOD_R = R<u8, TMOD_A>;

Reader of field TMOD

Implementations

impl TMOD_R

pub fn variant(&self) -> TMOD_A

Get enumerated values variant

pub fn is_tx_and_rx(&self) -> bool

Checks if the value of the field is TX_AND_RX

pub fn is_tx_only(&self) -> bool

Checks if the value of the field is TX_ONLY

pub fn is_rx_only(&self) -> bool

Checks if the value of the field is RX_ONLY

pub fn is_eeprom_read(&self) -> bool

Checks if the value of the field is EEPROM_READ