Module rp2040::xosc

Controls the crystal oscillator

Modules

count
A down counter running at the xosc frequency which counts to zero and stops.\n To start the counter write a non-zero value.\n Can be used for short software pauses when setting up time sensitive hardware.

ctrl
Crystal Oscillator Control

dormant
Crystal Oscillator pause control\n This is used to save power by pausing the XOSC\n On power-up this field is initialised to WAKE\n An invalid write will also select WAKE\n WARNING: stop the PLLs before selecting dormant mode\n WARNING: setup the irq before selecting dormant mode

startup
Controls the startup delay

status
Crystal Oscillator Status

Structs

RegisterBlock
Register block

Type Definitions

COUNT
A down counter running at the xosc frequency which counts to zero and stops.\n To start the counter write a non-zero value.\n Can be used for short software pauses when setting up time sensitive hardware.

CTRL
Crystal Oscillator Control

DORMANT
Crystal Oscillator pause control\n This is used to save power by pausing the XOSC\n On power-up this field is initialised to WAKE\n An invalid write will also select WAKE\n WARNING: stop the PLLs before selecting dormant mode\n WARNING: setup the irq before selecting dormant mode

STARTUP
Controls the startup delay

STATUS
Crystal Oscillator Status