Type Definition rp2040::xosc::ctrl::R

type R = R<u32, CTRL>;

Reader of register CTRL

Implementations

impl R

pub fn enable(&self) -> ENABLE_R

Bits 12:23 - On power-up this field is initialised to DISABLE and the chip runs from the ROSC.\n If the chip has subsequently been programmed to run from the XOSC then setting this field to DISABLE may lock-up the chip. If this is a concern then run the clk_ref from the ROSC and enable the clk_sys RESUS feature.\n The 12-bit code is intended to give some protection against accidental writes. An invalid setting will enable the oscillator.

pub fn freq_range(&self) -> FREQ_RANGE_R

Bits 0:11 - Frequency range. This resets to 0xAA0 and cannot be changed.